% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_MDS_plot.R
\name{applyMDSPlot}
\alias{applyMDSPlot}
\title{Correlation analysis}
\usage{
applyMDSPlot(
  data = data,
  plotmds = plotmds,
  databatch = databatch,
  data.batch = data.batch,
  group = group,
  ids = ids,
  prefix = prefix,
  technology = technology
)
}
\arguments{
\item{my.d1}{first dataframe with values to be correlated. These must have the same dimensions and rownames must the same.}

\item{my.d2}{second dataframe with values to be correlated. These must have the same dimensions and rownames must the same.}

\item{my.filename}{a name of output plot}
}
\value{
correlation plots
}
\description{
If two datasets are provided (for correlation and/or network analysis), this option should be specified as a comma separated list (without quotes or parenthesis!) of length two,
first entry referring to data file 1 and second entry referring to the data file 2.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
