% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_normalization.R
\name{applyNormalization}
\alias{applyNormalization}
\title{Normalization, Filtering and Transformation}
\usage{
applyNormalization(
  data,
  sdata = NULL,
  data.original,
  sdata.original = NULL,
  group,
  sgroup = NULL,
  transform,
  standardize,
  technology
)
}
\arguments{
\item{data}{a dataframe with expression/abundance counts}

\item{sdata}{a dataframe with expression/abundance counts for second dataset}

\item{data.original}{a original data}

\item{sdata.original}{a original data for second dataset}

\item{group}{a dataframe with samples groups}

\item{sgroup}{a dataframe with samples groups for second dataset}

\item{transform}{a transformation type}

\item{standardize}{a standardization type}

\item{technology}{a technology}
}
\value{
normalized, filtered and transformed data
}
\description{
Function for Normalization, Filtering and Transformation. This pipeline does not handle background correction of single-channel intensity data or within array normalization two-color intensity data. See limma manual section on array normalization for more on this. Data may be fully normalized with limma (R/Rstudio) or another software and the pipeline re-run.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
