% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_correction.R
\name{batchCorrect}
\alias{batchCorrect}
\title{Batch correction}
\usage{
batchCorrect(
  data,
  sdata = NULL,
  batch = NULL,
  sbatch = NULL,
  databatch = NULL,
  sdatabatch = NULL,
  group,
  sgroup = NULL,
  technology
)
}
\value{
batch corrected gene counts
}
\description{
Function for batch correction
}
\examples{
\dontrun{
...
}
}
\seealso{

}
