% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1}
\alias{campp2_brca_1}
\title{campp2_brca_1 - example data}
\format{
A data frame with 10000 rows (genes) and 30 columns (samples):
\describe{
}
}
\usage{
campp2_brca_1
}
\description{
A campp2_brca_1 contains 10000 gene counts for 30 samples: 20 tumor (4 subtypes, each subtype has 5 samples) x 10 normal samples). Data are derived from TCGA dataset.
}
\keyword{datasets}
