% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_batchCorrected}
\alias{campp2_brca_1_batchCorrected}
\title{campp2_brca_1_batchCorrected - example data}
\format{
A matrix with 8919 rows (genes) and 8 columns (samples).
}
\usage{
data(campp2_brca_1_batchCorrected)
}
\value{
A matrix with 8919 rows (genes) and 8 columns (samples).
}
\description{
A dataset with batch corrected data generated using "BatchCorrect" function.
This dataset is based on "campp2_brca_1_normalized". For more details, see the
vignette.
}
\keyword{datasets}
