% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_DEA}
\alias{campp2_brca_1_DEA}
\title{campp2_brca_1_DEA - example data}
\format{
a list of:
1) a data frame (311 x 8) - re-formatted matrix of differential
expression/abundance results from limma.
2) a list of original results (a list) from limma
3) a character vector with unique feature names
#' \describe{
}
}
\usage{
campp2_brca_1_DEA
}
\description{
A results from differential gene expression analysis provided by
RunDEA function using the provided example script. For more details, see the
vignette.
}
\keyword{datasets}
