% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_DEA_HUGO}
\alias{campp2_brca_1_DEA_HUGO}
\title{campp2_brca_1_DEA_HUGO - example data}
\format{
a data frame (43 x 9)
}
\usage{
data(campp2_brca_1_DEA_HUGO)
}
\value{
a data frame (43 x 9)
}
\description{
results from differential gene expression analysis provided by
RunDEA, filtering on group comparisons and applying AddGeneName:
step1:
campp2_brca_1_DEA<-RunDEA(data=campp2_brca_1, metadata=campp2_brca_1_meta,
group=campp2_brca_1_meta$subtype, prefix="test",
block=NULL, batch=campp2_brca_1_meta$age,
covarDEA = c("tumor_stage"), cutoff.logFC=1, cutoff.FDR=0.05)
step2:
campp2_brca_1_DEA_HUGO<-subset(campp2_brca_1_DEA, grepl("healthy",
comparison, fixed = TRUE))
step3:
campp2_brca_1_DEA_HUGO<-AddGeneName(campp2_brca_1_DEA_HUGO,ensembl.version)
}
\keyword{datasets}
