% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_DEA_HUGO_features_per_group}
\alias{campp2_brca_1_DEA_HUGO_features_per_group}
\title{campp2_brca_1_DEA_HUGO_features_per_group - example data}
\format{
a list of 4 character vectors (1341, 395, 55, 263)
\describe{
}
}
\usage{
campp2_brca_1_DEA_HUGO_features_per_group
}
\description{
This object represents a list of the features characteristic for each group
and was created based on campp2_brca_1_DEA_HUGO (described above) following
these steps:
control.group="healthy"
groups.full.list <- split.data.frame(campp2_brca_1_DEA_HUGO, campp2_brca_1_DEA_HUGO$comparison)
campp2_brca_1_DEA_HUGO_features_per_group=list()
for (features in groups.full.list){
    campp2_brca_1_DEA_HUGO_features_per_group <- append(campp2_brca_1_DEA_HUGO_features_per_group,list(features$name))
}
names(campp2_brca_1_DEA_HUGO_features_per_group) <- gsub("-","",gsub(control.group,"",names(groups.full.list)))
}
\keyword{datasets}
