% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_DEA_LASSO_consensus}
\alias{campp2_brca_1_DEA_LASSO_consensus}
\title{campp2_brca_1_DEA_LASSO_consensus - example data}
\format{
a data frame (7x8)
}
\usage{
data(campp2_brca_1_DEA_LASSO_consensus)
}
\value{
a data frame (7x8)
}
\description{
This object represents a data frame output from
campp2_brca_1_DEA_LASSO_consensus function. The table includes intersect
of the DEA and EN/LASSO/Ridge regression results. The format of the data
corresponds to the DEA (limma) output.
Data were generated by running:
##here we prepare DEA results compatible (group="diagnosis) with
EN/LASS/Ridge regression results:
campp2_brca_1_DEA_diagnosis<-RunDEA(data=campp2_brca_1_normalized,
metadata=campp2_brca_1_meta,
group=campp2_brca_1_meta$diagnosis, prefix="test", batch=campp2_brca_1_meta$age,
covarDEA = c("tumor_stage"), cutoff.logFC=1, cutoff.FDR=0.05)
##here we run RunDEA_LASSO_consensus function:
campp2_brca_1_DEA_LASSO_consensus <- RunDEA_LASSO_consensus(
campp2_brca_1_DEA_diagnosis$DEA.out,
campp2_brca_1_LASSO, group=campp2_brca_1_meta$diagnosis,
viridis.palette="turbo", "test")
}
\keyword{datasets}
