% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_DEA_out}
\alias{campp2_brca_1_DEA_out}
\title{campp2_brca_1_DEA_out - example data}
\format{
a list of:
1) a data frame (43 x 8) - re-formatted matrix of differential
expression/abundance results from limma.
2) a list of original results (a list) from limma
3) a character vector with unique feature names
4) a design matrix
5) a contrast matrix
}
\usage{
data(campp2_brca_1_DEA_out)
}
\value{
a list of:
1) a data frame (43 x 8) - re-formatted matrix of differential
expression/abundance results from limma.
2) a list of original results (a list) from limma
3) a character vector with unique feature names
4) a design matrix
5) a contrast matrix
}
\description{
results from differential gene expression analysis provided by
ExportDEA function using the provided example script:
campp2_brca_1_DEA_out<-ExportDEA(res.DEA = campp2_brca_1_DEA$res.DEA,
                                 prefix="test")
}
\keyword{datasets}
