% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_distributionsFit}
\alias{campp2_brca_1_distributionsFit}
\title{campp2_brca_1_distributionsFit - example data}
\format{
A list of 10 lists describing fitting the normal distribution.
\describe{
}
}
\usage{
campp2_brca_1_distributionsFit
}
\description{
A results from FitDistributions function providing a list of distribution
descriptions for first 10 features.
}
\keyword{datasets}
