% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_LASSO}
\alias{campp2_brca_1_LASSO}
\title{campp2_brca_1_LASSO - example data}
\format{
a list of dataframe (42x1) and 2 empty variables
\describe{
}
}
\usage{
campp2_brca_1_LASSO
}
\description{
This object represents an output from runLASSO function. The list
includes a dataframe of feature names detected by both DEA and EN/LASSO/Ridge
regression. Another 2 sub-objects from the list for cross validation error
and roc. are empty because the input data (campp2_brca_1_batchCorrected)
is not large enough for the computation of those metrics.
Data were generated by running:
campp2_brca_1_LASSO<-runLASSO(data=campp2_brca_1_batchCorrected,
group=campp2_brca_1_meta$diagnosis, alpha=0.5, min.coef=0, nfolds=10,
prefix="test")
}
\keyword{datasets}
