% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_meta}
\alias{campp2_brca_1_meta}
\title{campp2_brca_1_meta - example data}
\format{
A data frame with 30 rows (samples) and 10 columns.
\describe{
}
}
\usage{
campp2_brca_1_meta
}
\description{
A metadata for campp2_brca_1 contains 9 variables for 30 samples in
campp2_brca_1 dataset. Data are derived from TCGA dataset.
}
\keyword{datasets}
