% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_NAs}
\alias{campp2_brca_1_NAs}
\title{campp2_brca_1_NAs - example data}
\format{
A matrix with 10000 rows (genes) and 8 columns (samples).
}
\usage{
data(campp2_brca_1_NAs)
}
\value{
A matrix with 10000 rows (genes) and 8 columns (samples).
}
\description{
A dataset with randomly introduced 1000 NA values into every sample's genes
(10000). This dataset is based on "campp2_brca_1". For more details, see the
vignette.
}
\keyword{datasets}
