% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_normalized}
\alias{campp2_brca_1_normalized}
\title{campp2_brca_1_normalized - example data}
\format{
A Elist with 9046 rows (genes) and 30 columns (samples).
\describe{
}
}
\usage{
campp2_brca_1_normalized
}
\description{
A dataset with normalized and transformed gene counts data generated using
"NormalizeData" function. This dataset is based on "campp2_brca_1_zeroFix" data.
For more details, see the vignette.
}
\keyword{datasets}
