% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_run_rf}
\alias{campp2_brca_1_run_rf}
\title{campp2_brca_1_run_rf.rda - example data}
\format{
a list containing two elements
}
\usage{
data(campp2_brca_1_run_rf)
}
\value{
a list containing two elements
}
\description{
This object represents output from the
RunRF function which is used to run the RFApply function.
Data were generated by running:
campp2_brca_1_run_rf <-
RunRF(data = campp2_brca_1_batchCorrected,
group = campp2_brca_1_meta$diagnosis,
split.size = 5,
test.train.ratio = 0.25,
num.trees.init = 5000,
num.trees.iterat = 2000)
}
\keyword{datasets}
