% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_2}
\alias{campp2_brca_2}
\title{campp2_brca_2 - example data}
\format{
A data frame with 10000 rows (genes) and 25 columns (samples):
\describe{
}
}
\usage{
campp2_brca_2
}
\description{
A campp2_brca_2 containing 10000 gene counts for 25 samples (mix of tumor x normal samples). Data are derived from TCGA dataset.
}
\keyword{datasets}
