% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_2_meta}
\alias{campp2_brca_2_meta}
\title{campp2_brca_2_meta - example data}
\format{
A data frame with 8 rows (samples) and 10 columns.
\describe{
}
}
\usage{
campp2_brca_2_meta
}
\description{
A metadata for campp2_brca_2 contains 9 variables for 8 samples in
campp2_brca_2 dataset. Data are derived from TCGA dataset.
}
\keyword{datasets}
