% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_check.R
\name{checkDistr}
\alias{checkDistr}
\title{Distribution check}
\usage{
checkDistr(
  data,
  sdata = NULL,
  datacheck,
  databatch,
  sdatabatch,
  data.batch = NULL,
  sdata.batch = NULL
)
}
\arguments{
\item{data}{a gene counts matrix}

\item{sdata}{a gene counts matrix for a secend dataset}

\item{datacheck}{TRUE/FALSE parameter for distributional check}

\item{databatch}{TRUE/FALSE parameter for batch correction}

\item{sdatabatch}{TRUE/FALSE parameter for batch correction for the second dataset}

\item{data.batch}{batch corrected gene counts}

\item{sdata.batch}{batch corrected gene counts for the second dataset}
}
\value{
batch corrected gene counts
}
\description{
Function for checking gene counts distribution
}
\examples{
\dontrun{
...
}
}
\seealso{

}
