% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_zeros.R
\name{checkZeros}
\alias{checkZeros}
\title{Detect and replace zero values in the data}
\usage{
checkZeros(data, sdata = NULL)
}
\arguments{
\item{my.data}{a dataframe of expression/abundance counts, N.B only a subset of variables should be input, not intended for the full expression matrix!}
}
\value{
normalized counts
}
\description{
Checking the presence of zero values in the data and replacing them.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
