% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_plot.R
\name{chunk2}
\alias{chunk2}
\title{Survival plot}
\usage{
chunk2(x, n)
}
\arguments{
\item{my.survres}{a survival object in the form of a list with a cox regression for each feature.}

\item{my.filename}{a name of output plot}

\item{my.n}{?}
}
\value{
survival plot
}
\description{
Function for generating survival plot
}
\examples{
\dontrun{
...
}
}
\seealso{

}
