% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_analysis.R
\name{CorrAnalysis}
\alias{CorrAnalysis}
\title{Correlation analysis}
\usage{
CorrAnalysis(d1, d2, my.filename)
}
\arguments{
\item{my.filename}{a name of output plot}

\item{my.d1}{first dataframe with values to be correlated. These must have the same dimensions and rownames must the same.}

\item{my.d2}{second dataframe with values to be correlated. These must have the same dimensions and rownames must the same.}
}
\value{
correlation plots
}
\description{
If two datasets are provided (for correlation and/or network analysis), this option should be specified as a comma separated list (without quotes or parenthesis!) of length two,
first entry referring to data file 1 and second entry referring to the data file 2.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
