% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_plots.R
\name{CorrelationPlots}
\alias{CorrelationPlots}
\title{Correlation plots}
\usage{
CorrelationPlots(my.data, my.serumdata, my.features, my.filename)
}
\arguments{
\item{my.data}{= a dataframe with expression/abundance counts for tissue or TIF}

\item{my.serumdata}{= a dataframe with expression/abundance counts for serum}

\item{my.features}{?}

\item{my.filename}{= name of output plot}
}
\value{
correlation plots
}
\description{
Function for correlation scatter plots
}
\examples{
\dontrun{
...
}
}
\seealso{

}
