% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_Feature_apply.R
\name{DAFeatureApply}
\alias{DAFeatureApply}
\title{Apply DAFeature}
\usage{
DAFeatureApply(
  my.contrasts,
  my.data,
  my.design,
  cutoff.logFC,
  cutoff.FDR,
  my.block = NULL,
  my.vector = FALSE
)
}
\arguments{
\item{my.contrasts}{= a contrast between groups of interest}

\item{my.data}{a dataframe of expression/abundance counts}

\item{my.design}{a design matrix with all comparisons}

\item{cutoff.logFC}{a cutoff for logFC}

\item{cutoff.FDR}{a cutoff for FDR}

\item{my.block}{if blocking than a vector of patient IDs}

\item{my.vector}{a vector of patient IDs; TRUE/FALSE statement specifying output format, if TRUE the function return a vector of feature IDs only}
}
\value{
DA/DE features
}
\description{
Apply differential abundance analysis to all the comparisons at once
}
\examples{
\dontrun{
...
}
}
