% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_Feature_apply.R
\name{DEAFeatureApply}
\alias{DEAFeatureApply}
\title{Apply DEA_Feature}
\usage{
DEAFeatureApply(
  contrast.matrix,
  data,
  design,
  cutoff.logFC,
  cutoff.FDR,
  block,
  vector = FALSE
)
}
\arguments{
\item{contrast.matrix}{an array containing contrast.matrixs between groups of interest.}

\item{data}{A raw gene count matrix from seq, array, ms or other technology (with gene IDs as row names and sample IDs as columns). It's recommended to import gene counts using function "import_counts".}

\item{design}{a design matrix with all comparisons.}

\item{cutoff.logFC}{a number specifying the cutoff for LogFC.}

\item{cutoff.FDR}{a number specifying the cutoff for FDR.}

\item{block}{A vector or factor specifying a blocking variable. The block must be of same length as data and contain 2 or more options. For 2 datasets, the block can be defined as a vector of the two seperate blocks.}

\item{vector}{a vector of patient IDs; TRUE/FALSE statement specifying output format. If TRUE the function return a vector of feature IDs only.}
}
\value{
DEA features for all comparisons
}
\description{
A function for applying differential expression/abundance analysis to all comparisons at once.
}
\examples{
\dontrun{
...
}
}
