% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{Design.Matrix}
\alias{Design.Matrix}
\title{Create model matrix}
\usage{
Design.Matrix(metadata, group, batch = NULL, covariates = NULL)
}
\arguments{
\item{metadata}{Samples' metadata table should be imported using function
"import_metadata". Metadata must include exactly the same samples as gene
counts (data1) and samples must be sorted similarly.}

\item{group}{a factor specifying group for each sample (e.g. could be
represented by a column from a metadata file)}

\item{batch}{a factor specifying batch for each sample (e.g. could be
represented by a column from a metadata file). Default = NULL.}

\item{covariates}{a covariate(s) to include in the analysis provided
as a character vector. covarDEA can be defined only with a batch parameter}
}
\value{
a model matrix used e.g. for limma
}
\description{
A function for creating a model matrix based on gene counts
metadata table and selected column names (covariates) from the metadata.
}
\examples{
{
design.matrix<-Design.Matrix(campp2_brca_1_meta, campp2_brca_1_meta$diagnosis)
}
}
