% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF2List.R, R/Subset_DA.R
\name{DFtoList}
\alias{DFtoList}
\title{Dataframe to List}
\usage{
DFtoList(my.data)

DFtoList(my.data)
}
\arguments{
\item{my.data}{a dataframe with results of differential expression analysis.}
}
\value{


a list of dataframes by comparison
}
\description{
Function which subsets a dataframe of differential expression analysis into a list of dataframes by comparison.

Function which subsets a dataframe of differential expression analysis into a list of dataframes by comparison.
}
\examples{
\dontrun{
...
}
\dontrun{
...
}
}
\seealso{



}
