% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_STRING.R, R/STRING.R
\name{DownloadPPInt}
\alias{DownloadPPInt}
\title{Download string database}
\usage{
DownloadPPInt(my.geneIDs, my.version = "11.0")

DownloadPPInt(my.geneIDs, my.version = "11.0")
}
\arguments{
\item{my.geneIDs}{= String specifying type of gene ID to use to match IDs in STRING database. Options are: ensembl_peptide_id, hgnc_symbol, ensembl_gene_id, ensembl_transcript_id or uniprotswissprot.}

\item{my.version}{= Version of STRING database to use. Default is version 11.}

\item{my.DB}{= output of the function "DownloadPPInt", e.g. a curated string database.}

\item{my.Geneset}{= a dataframe with results of differential expression analysis.}
}
\value{



}
\description{
Function to download string database

FUNCTION TO OBTAIN STRING DATABASE AND ANALYZE PROTEIN-PROTEIN INTERACTIONS
}
\examples{
\dontrun{
...
}
\dontrun{
...
}
}
\seealso{



}
