% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_DEA.R
\name{ExportDEA}
\alias{ExportDEA}
\title{Export DEA results into .txt}
\usage{
ExportDEA(res.DEA, filename)
}
\arguments{
\item{res.DEA}{A list of data frames from DEA_feature_apply containing gene counts, p-values, FDRs and logFCs.}

\item{filename}{The name of the output file.}
}
\value{
DEA results as .txt table
}
\description{
A function for exporting DEA results into a .txt file.
}
\examples{
\dontrun{
...
}
}
