% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tumor_normal.R
\name{extract_tumor_normal_samples}
\alias{extract_tumor_normal_samples}
\title{RegExp extraction of gene counts}
\usage{
extract_tumor_normal_samples(
  gene_count_matrix,
  normal_sampleRegExp = "....-..-....-1[1-9].-...-....-..",
  tumor_sampleRegExp = "....-..-....-0[1-9].-...-....-.."
)
}
\arguments{
\item{gene_count_matrix}{gene count matrix to be processed}

\item{normal_sampleRegExp}{regular expression to extract normal samples (based on the sample name); default value fits TCGA data (https://docs.gdc.cancer.gov/Encyclopedia/pages/TCGA_Barcode/)}

\item{tumor_sampleRegExp}{regular expression to extract tumor samples}
}
\value{
As an output, list of 1) read counts from paired normal samples ($normal_paired), 2) read counts from paired tumor samples ($tumor_paired), 3) counts from all normal samples ($normal), 4)  counts from all tumor samples ($tumor), 5) sample names of duplicated normal samples ($N_dup_names), 6) sample names of duplicated tumor samples ($T_dup_names), 7) sample names from all normal samples ($all_normal_names), 8) sample names from all tumor samples ($all_tumor_names) is generated.
}
\description{
This function extracts gene counts based on sample name using regular expressions. This might be useful e.g. when working with TCGA datasets.
}
\examples{
\dontrun{
    load("~/opt/campp_bioconducor/CAMPP2/data/BRCA/BRCA_dataPrep.rda")
    normal_tumor_counts<-extract_tumor_normal_samples(dataPrep)
    normal_counts_paired<-as.data.frame(normal_tumor_counts$normal_paired)
    tumor_counts_paired<-as.data.frame(normal_tumor_counts$tumor_paired) #including technical replicates
    normal_counts<-as.data.frame(normal_tumor_counts$normal)
    tumor_counts<-as.data.frame(normal_tumor_counts$tumor)
    normal_duplicates<-normal_tumor_counts$N_dup_names
    tumor_duplicates<-normal_tumor_counts$T_dup_names
    all_normal_names<-normal_tumor_counts$all_normal_names
    all_tumor_names<-normal_tumor_counts$all_tumor_names
    write.table(tumor_counts_paired, file="tumor_counts_paired.txt", sep = "\t")
    write.table(normal_counts_paired, file="normal_counts_paired.txt", sep = "\t")
    write.table(tumor_counts, file="tumor_counts.txt", sep = "\t")
    write.table(normal_counts, file="normal_counts.txt", sep = "\t")
    write.table(tumor_duplicates, file="tumor_duplicates.txt", sep = "\t")
    write.table(normal_duplicates, file="normal_duplicates.txt", sep = "\t")
}
}
