% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting_data_distribution.R
\name{FitDistributions}
\alias{FitDistributions}
\title{Fitting Data Distributions}
\usage{
FitDistributions(data)
}
\arguments{
\item{data}{a data frame of expression/abundance counts, N.B only a subset of
variables should be input, not intended for the full expression matrix!
Depending on the character of the data, various distributions are checked.
In case of only integer numbers present in the dataset, poison and normal
distributions are tested. In case of presence of float values in the dataset,
weibull, gamma, log normal and normal distributions are tested. In case of
negative float values, only normal distribution is calculated.}
}
\value{
a list of the results from fitdist function
}
\description{
A function for fitting data distribution
}
\examples{
\dontrun{
campp2_brca_1_distributions <- FitDistributions(campp2_brca_1[1:10,])
}
}
