% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_zeros.R
\name{FixZeros}
\alias{FixZeros}
\title{Fix zeros}
\usage{
FixZeros(data, group, remove.sparse.features = TRUE)
}
\arguments{
\item{data}{a dataframe of gene/abundance counts.}

\item{group}{a factor specifying group which should be represented by
a column from a metadata file.}

\item{remove.sparse.features}{a logical argument (TRUE/FALSE) for removal of
features with sum of zero counts larger than the size of the smallest
sample group. Default is TRUE.}
}
\value{
a data frame with fixed zeros. Features having sum of zero counts
        higher than the size of the smallest sample group are removed by
        default.
}
\description{
A function for a detection of zero/negative values and for
             replacing zero values in the data. Features having sum of zero
             counts higher than the size of the smallest sample group will be
             removed by default. Zeros are substituted with minimal values
             (>0) observed for each feature.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
