% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_Features.R
\name{Forest_Features}
\alias{Forest_Features}
\title{Random forest fitting and variable selection}
\usage{
Forest_Features(
  seed,
  data,
  group,
  validation = FALSE,
  num_trees,
  num_trees_iterat
)
}
\arguments{
\item{seed}{an integer containing a random seed number.}

\item{data}{a matrix of (transformed and normalized) feature counts from
"seq", "array", "ms" or "other" technology (with feature IDs as row names
and sample IDs as columns).}

\item{group}{a vector of type factor containing the group ID for each
sample}

\item{validation}{a boolean indicating if validation will be performed
on test data. If TRUE validation will be performed on test data. If FALSE
validation will not be performed on test data. Defaults to FALSE.}

\item{num_trees}{an integer specifying number of trees to use for the first
forest in the feature selection process}

\item{num_trees_iterat}{an integer specifying number of trees to use for
all additional forests in the feature selection process}
}
\value{
a list of four elements: 1) results of variable selection using random forest,
2) random forest model fitted to data, 3) predictions of test data using fitted random
forest model and 4) confusion matrix of test data using fitted random forest model.
If my.validation = FALSE, the last three elements in output will be NA.
}
\description{
This function implements random forest on feature counts
(e.g. genes) and allows for feature selection using the random forest
algorithm.
}
\examples{
\dontrun{
campp2_brca_1_forest_features <-
Forest_Features(seed = 173,
data = campp2_brca_1,
group = campp2_brca_1_meta$diagnosis,
validation = TRUE,
num_trees = 30,
num_trees_iterat = 15)
}
}
