% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_PPI.R, R/PPI.R
\name{GetDeaPPInt}
\alias{GetDeaPPInt}
\title{Extract the protein-protein interactions}
\usage{
GetDeaPPInt(my.DB, my.Genedat)

GetDeaPPInt(my.DB, my.Genedat)
}
\arguments{
\item{my.DB}{= An output of the function "DownloadPPInt", e.g. a curated string database.}

\item{my.Geneset}{= A dataframe with results of differential expression analysis.}
}
\value{



}
\description{
Function which extracts the protein-protein interactions where each protein (gene) is differentially abundant (expressed).

A Function which extracts the protein-protein interactions where each protein (gene) is differentially abundant (expressed).
}
\examples{
\dontrun{
...
}
\dontrun{
...
}
}
\seealso{



}
