% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_gene_miRNA_interaction.R
\name{GetGeneMiRNAInt}
\alias{GetGeneMiRNAInt}
\title{GetGeneMiRNAInt}
\usage{
GetGeneMiRNAInt(arg.miRset, my.miRdat, my.Genedat = NULL)
}
\arguments{
\item{arg.miRset}{a string specifying what miRNA database to use, options are: mirtarbase (validated), targetscan (predicted) or tarscanbase (validated and predicted).}

\item{my.miRdat}{a dataframe with results of differential expression analysis (miRNAs).}

\item{my.Genedat}{?}
}
\value{

}
\description{
A function which extracts the miRNA-gene interactions where miRNAs (and potentially genes, if this data is available,) are differentially expressed.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
