% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_counts.R
\name{importCounts}
\alias{importCounts}
\title{Read gene counts}
\usage{
importCounts(my.data)
}
\arguments{
\item{my.data}{a dataframe of expression/abundance counts.}
}
\description{
Reading gene count matrix files in txt or xlsx format. Check a sample name format after the import.
}
\examples{
\dontrun{
...
}
}
