% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_metadata.R
\name{importMetadata}
\alias{importMetadata}
\title{Read metadata}
\usage{
importMetadata(my.data)
}
\arguments{
\item{my.data}{a dataframe of metadata table.}
}
\description{
Reading metadata file in txt or xlsx format
}
\examples{
\dontrun{
...
}
}
