% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASSO.R
\name{LASSOFeature}
\alias{LASSOFeature}
\title{LASSO function}
\usage{
LASSOFeature(
  my.seed,
  my.data,
  my.group,
  my.LAorEN,
  my.validation = FALSE,
  my.multinorm = TRUE
)
}
\arguments{
\item{my.seed}{a randomly generated vector of seeds}

\item{my.data}{a data marix of values}

\item{my.group}{a vector of integers specifying group}

\item{my.LAorEN}{a hyperparameter alpha. This value must be set to 0.0 < x < 1.0 for Elastic Net (0.5 is default) or to 1.0 for LASSO regression.}

\item{my.validation}{perform validation TRUE/FALSE}

\item{my.multinorm}{perfom multinorm TRUE/FALSE; IF my.multinorm=TRUE, then analysis will be multinomial, IF my.multinorm=FALSE, then then analysis will be binomial}
}
\value{
DA/DE features from LASSO
}
\description{
Run LASSO function to obtain DA features
}
\examples{
\dontrun{
...
}
}
