% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{MakeHeatmap}
\alias{MakeHeatmap}
\title{Make Heatmap}
\usage{
MakeHeatmap(data, groups, prefix, gradient = viridis(300, option = "cividis"))
}
\arguments{
\item{data}{A dataframe containing expression values for genes within a set of samples.}

\item{prefix}{A prefix for the output filename.}

\item{gradient}{A color gradient to use for the heatmap. Default is 'blue' using the viridis library.}

\item{group}{a factor derived from metadata column selected as a sample group (e.g. diagnosis), which will be used to color and split the heatmap by. Length should be equal to ncol(data)).}
}
\value{
Heatmap
}
\description{
A function for making heatmaps to showcase the difference in expression values across samples.
}
\examples{
\dontrun{
...
}
}
