% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpsetPlot.R
\name{MakeUpset}
\alias{MakeUpset}
\title{Upset Plot}
\usage{
MakeUpset(prefix, groups.feature.list)
}
\arguments{
\item{prefix}{A prefix for the output filename.}

\item{groups.feature.list}{A list of character vectors (each character
vector named by its group name, e.g., "LumA") including feature names
characteristic for each group of interest (e.g., subtype).}
}
\value{
an Upset plot showcasing the size of intersections of the features
between the sample groups
}
\description{
A function for making Upset plots to visualize the size of
intersections of the features present each group
(e.g., differentially expressed genes for each subtype)
}
\examples{
\dontrun{
MakeUpset("testUpset",campp2_brca_1_DEA_HUGO_features_per_group)
}
}
