% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennDiagram.R
\name{MakeVennDiagram}
\alias{MakeVennDiagram}
\title{Venn Diagram}
\usage{
MakeVennDiagram(sets_list, names_sets, prefix)
}
\arguments{
\item{sets_list}{A list of sets to compare.}

\item{names_sets}{A vector of the corresponding names to each set in sets_list.}

\item{prefix}{A prefix for the output filename.}
}
\value{
Venn Diagram
}
\description{
A function for making Venn Diagrams to visualize the intersection between 2 or more sets.
}
\examples{
\dontrun{
...
}
}
