% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match_PP_GmiR.R
\name{MatchPPGmiRInt}
\alias{MatchPPGmiRInt}
\title{Match PP GmiR Intactions}
\usage{
MatchPPGmiRInt(nodes.list1, nodes.list2)
}
\arguments{
\item{nodes.list1}{a list of networks, i.e. output of the function "GetGeneMiRNAInt"}

\item{nodes.list2}{a list of networks, i.e. output of the function "GetDeaPPInt"}
}
\value{

}
\description{
A function for combining miRNA-gene interactions and protein-protein (gene-gene) interactions.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
