% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_plot.R
\name{MDSPlot}
\alias{MDSPlot}
\title{Multidimensional Scaling Plot}
\usage{
MDSPlot(data, group, labels, cols, MDS.labels)
}
\arguments{
\item{data}{a dataframe of expression/abundance counts}

\item{group}{a vector of integers specifying the group (may be the same or different, length should be equal to ncol(dataframe))}

\item{cols}{a vector of colors (one color for each group)}

\item{MDS.labels}{a TRUE/FALSE statement specifying whether to include labels on the output plot}

\item{lables}{a vector of IDs for labeling (may be the same or different, length should be equal to ncol(dataframe))}
}
\value{
distribution plots
}
\description{
Data overview. The plot shows the relationship between samples (squared euclidian distance) in the dataset. Each dot represents one sample.
}
\examples{
\dontrun{
...
}
}
