% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCounts.R
\name{meanCounts}
\alias{meanCounts}
\title{Get mean and SD of the feature counts}
\usage{
meanCounts(data, group)
}
\arguments{
\item{data}{a data frame of feature counts.}

\item{group}{a factor specifying group for each sample (e.g. could be represented by a column from a metadata file)}
}
\value{
a list of data frames (for each sample group) describing mean and SD
for all sample groups and also for each group of samples. Results are
visualized via graphs saved into current directory as pdf files.
}
\description{
A function for obtaining an average feature counts and SD for
each gene across all the samples and for each sample group. Results are
provided in a form of a list of data frames and figures which are exported
into .pdf.
}
\examples{
{
meanCounts(campp2_brca_1_batchCorrected, campp2_brca_1_meta$diagnosis)
}
}
