% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_IC.R
\name{ModuleIC}
\alias{ModuleIC}
\title{ModuleIC}
\usage{
ModuleIC(
  vec.of.modulecolors,
  my.moduleColors,
  my.IC,
  my.ExpData,
  my.n,
  my.softPower,
  my.name
)
}
\arguments{
\item{vec.of.modulecolors}{A vector of module colors}

\item{my.moduleColors}{A module color object from WGCNA}

\item{my.IC}{A interconnectivity object from WGCNA}

\item{my.ExpData}{A dataset, features (genes, proteins ect) as columns and samples as rows.}

\item{my.n}{A Number indicating top n% most interconnected features in dataset}

\item{my.softPower}{..to describe..}

\item{my.name}{A name of output plot(s)}
}
\description{
Weighted gene co-expression network analysis for describing the correlation patterns among genes
}
\examples{
\dontrun{
...
}
}
