% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_data.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalization and Transformation}
\usage{
NormalizeData(data, data.original, group, transform, standardize, technology)
}
\arguments{
\item{data}{a dataframe with expression/abundance counts (genes as rows; samples as columns), N.B only a subset of variables should be input, not intended for the full expression matrix!}

\item{data.original}{an original data1 (data.frame) including zero values}

\item{group}{a dataframe with groups metadata for each sample (given from metadata table)}

\item{transform}{a string vector of length 1 (or two in case of 2 datasets) defining transformation type for each dataset ("log2", "logit", "log10" or voom).}

\item{standardize}{a string vector of length 1 (or two in case of 2 datasets) defining standardization method  ("mean" or "median"); By default, data from "seq" technology are normalized by "TMM"; "array" technology is standardized using "quantile"}

\item{technology}{a string vector of length 1 (or two in case of 2 datasets) defining technology used for generating the data. Allowed types are: 'array', 'seq', 'ms' or 'other'.}
}
\value{
Elist of normalized, filtered and transformed (gene) counts data
}
\description{
This function normalizes and transforms feature counts depending on data types and selected normalization/standardization methods.
}
\examples{
\dontrun{
...
}

}
\seealso{

}
