% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_distributions.R
\name{PlotDistributions}
\alias{PlotDistributions}
\title{Plotting Distributions}
\usage{
PlotDistributions(data, fitted.data)
}
\arguments{
\item{data}{a data.frame of expression/abundance counts. N.B only a subset of variables should be input, not intended for the full expression matrix!(See function "FitDistributions")}

\item{fitted.data}{a list of lists obtained as output from the function "FitDistributions". The data and fitted.data must have the same dimensions, e.g. length(fitted.data) == nrows(data).}
}
\value{
distribution plots
}
\description{
Plotting the counts' Distributions
}
\examples{
\dontrun{
...
}
}
