% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_K_means.R
\name{PlotKmeans}
\alias{PlotKmeans}
\title{Plot Kmeans}
\usage{
PlotKmeans(my.data, clus.list, k, my.labels, my.name)
}
\arguments{
\item{my.data}{a dataframe of expression/abundance counts}

\item{clus.list}{a list of k-means?}

\item{k}{A number of clusters tested will be based on number of samples, fewer samples will result in fewer kmeans tested.}

\item{my.labels}{metadata column for sample labeling}

\item{my.name}{?dataset name?}
}
\value{
clustering results and plots
}
\description{
Function for Plotting Kmeans clustering; A folder with MDS plots will be returned for the best n cluster, based on BIC
}
\examples{
\dontrun{
...
}
}
