% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_K_means.R
\name{PlotMDSKmeans}
\alias{PlotMDSKmeans}
\title{Plot Kmeans}
\usage{
PlotMDSKmeans(data, n.clusters.list, nks, labels.kmeans, prefix)
}
\arguments{
\item{data}{a dataframe of expression/abundance counts}

\item{n.clusters.list}{a list of the numbers of clusters estimated by BIC}

\item{nks}{A number of clusters tested will be based on number of samples, fewer samples will result in fewer kmeans tested.}

\item{labels.kmeans}{metadata column for sample labeling (correct)}

\item{prefix}{?dataset name?}
}
\value{
clustering results and plots
}
\description{
a function for Plotting Kmeans clustering. MDS
plots will be returned for the n clusters, n is based on BIC.
}
\examples{
\dontrun{
...
}
}
