% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_files.R
\name{ReadMyFile}
\alias{ReadMyFile}
\title{Read files}
\usage{
ReadMyFile(my.data, data.type)
}
\arguments{
\item{my.data}{a dataframe of expression/abundance counts.}

\item{data.type}{select "counts" or "metadata" type. (Originally, this param was TRUE or FALSE defined by CAMPP.R)}
}
\description{
Reading input files
}
\examples{
\dontrun{
...
}
}
