% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_NAs.R
\name{ReplaceNAs}
\alias{ReplaceNAs}
\title{Replace "NA" values}
\usage{
ReplaceNAs(data, pct.NA.row = 70, pct.NA.column = 80)
}
\arguments{
\item{data}{a dataframe of gene/abundance counts.}

\item{pct.NA.row}{a number defining maximal percentage of NA values per row (feature). Rows with a higher percentage of NA values will be removed (70 by default).}

\item{pct.NA.column}{a number defining maximal percentage of NA values per column (sample). Columns with a higher percentage of NA values will be removed (80 by default).}
}
\value{
Data frame with replaced "NA" values
}
\description{
The features (rows) and samples (columns) with high proportion of "NAs" will be removed based on defined thresholds (70% for rows and 80% for columns by default). Remaining "NA" values will be replaced by imputed values. Missing value imputation is done using llsImpute (1st round). If 1st round wasn't successful or resulted in negative values, impute.knn algorithm is applied (rows with more than 50% missing values will be imputed using the overall mean per sample).
}
\examples{
\dontrun{
...
}
}
