% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_zeros.R
\name{ReplaceZero}
\alias{ReplaceZero}
\title{Replace zeros}
\usage{
ReplaceZero(data, group)
}
\arguments{
\item{data}{a dataframe of expression/abundance counts.}

\item{group}{a factor specifying group; should be represented by a column in metadata file.}
}
\value{
Data frame with replaced zeros
}
\description{
Replaces zero values in the data (counts).
}
\examples{
\dontrun{
...
}
}
