% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_Apply.R
\name{RF_Apply}
\alias{RF_Apply}
\title{Applying random forest across seeds}
\usage{
RF_Apply(data, group, validation, num_trees, num_trees_iterat)
}
\arguments{
\item{data}{a matrix of (transformed and normalized) feature counts from
"seq", "array", "ms" or "other" technology (with feature IDs as row names
and sample IDs as columns).}

\item{group}{a vector of type factor containing the group ID for each
sample}

\item{validation}{a boolean indicating if validation will be performed
on test data. If TRUE validation will be performed on test data. If FALSE
validation will not be performed on test data. Defaults to FALSE.}

\item{num_trees}{an integer specifying number of trees to use for the first
forest in the feature selection process}

\item{num_trees_iterat}{an integer specifying number of trees to use for
all additional forests in the feature selection process}
}
\value{
a list of eight elements: 1) selected variables from random forest
feature selection process, 2) out-of-bag errors from random forest feature
selection process, 3) OOB error from first random forest model fitted during
feature selection process, 4) confusion matrix of predictions made based on
OOB data and first random forest model fitted during feature selection process,
5) importance of features from first random forest model fitted during feature
selection process, 6) OOB error from fitted random forest model,
7) accuracy and 95% confidence interval for predictions of test data using
fitted random forest model, and 8) seeds used in the procedure.
}
\description{
This function applies the Forest_Features function on 10 random
seeds. The Forest_Features function performs random forest on feature counts
(e.g. genes) and allows for feature selection using the random forest
algorithm.
}
\examples{
\dontrun{
campp2_brca_1_rf_apply <-
RF_Apply(data = campp2_brca_1,
group = campp2_brca_1_meta$diagnosis,
validation = TRUE,
num_trees = 30,
num_trees_iterat = 15)
}
}
