% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_Apply.R
\name{RFApply}
\alias{RFApply}
\title{Applying random forest across seeds}
\usage{
RFApply(
  data,
  group,
  validation = FALSE,
  test.train.ratio,
  num.trees.init = 5000,
  num.trees.iterat = 2000
)
}
\arguments{
\item{data}{a matrix of (transformed and normalized) feature counts from
"seq", "array", "ms" or "other" technology (with feature IDs as row names
and sample IDs as columns). If available, batch corrected data should be used.}

\item{group}{a factor specifying group for each sample (e.g. could be
represented by a column from a metadata file)}

\item{validation}{a boolean indicating if validation will be performed
on test data. If TRUE validation will be performed on test data. If FALSE
validation will not be performed on test data. Default is FALSE.}

\item{test.train.ratio}{a floating point number between 0 and 1 representing
the ratio of samples to keep as validation dataset. For example, a
test.train.ratio = 0.25 splits 25 percent of the data into a validation dataset,
meaning 75 percent of the data will be kept as the training dataset.}

\item{num.trees.init}{an integer specifying number of trees to use for the first
forest in the feature selection process. Default is 5000.}

\item{num.trees.iterat}{an integer specifying number of trees to use for
all additional forests in the feature selection process. Default is 2000.}
}
\value{
a list of seven elements: 1) a list where each element is a character string
containing selected variables from the random forest feature selection process from
each seed run, 2) a list where each element is out-of-bag errors (of type numeric)
from all random forest models used in the feature selection process for each seed run,
3) a list where each element is the OOB error (of type numeric) from the best random
forest model (i.e. the final selected model) of the feature selection process for each
seed run, 4) a matrix containing average initial importance (before any variable deletion)
for each feature across seed runs, 5) a list where each element is the OOB error (of type
numeric) from fitted random forest model for each seed run (if a random forest was not
fitted then this will be NA), 6) a matrix ccontaining accuracy and 95 percent confidence
interval for predictions of test data using fitted random forest model where each row
in the matrix is a seed run, and 7) a vector of integers containing seeds used in the
procedure.
}
\description{
This function applies the ForestFeatures function using 10 random
seeds. The ForestFeatures function performs random forest on feature counts
(e.g. genes) and allows for feature selection using the random forest
algorithm. This function extracts various results from the feature selection
process such as the selected variables and the OOB errors. Moreover, the OOB
error which resulted in the final selected model and the final selected
variables is extracted. The final selected model is the one with the smallest
number of genes whose error rate is within 1 standard error of the minimum
error rate of all forests. The function also extracts results from the process
of random forest fitting to the input data such as OOB error, predictions on
test data and corresponding accuracies. Finally, the function calculates the
average initial importance of each variable across the seed runs.
}
\examples{
\dontrun{
campp2_brca_1_batchCorrected<-BatchCorrect(data=campp2_brca_1_normalized,
batch=campp2_brca_1_meta$tumor_stage,group=campp2_brca_1_meta$diagnosis,
technology="seq")

campp2_brca_1_rf_apply <-
RFApply(data = campp2_brca_1_batchCorrected,
group = campp2_brca_1_meta$diagnosis,
validation = TRUE,
test.train.ratio = 0.25,
num.trees.init = 5000,
num.trees.iterat = 2000)
}
}
