% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDEAVisuals.R
\name{RunDEAVisuals}
\alias{RunDEAVisuals}
\title{A wrapper for a visualisation of differential expression/abundance
analysis}
\usage{
RunDEAVisuals(
  data,
  cutoff.FDR = 0.01,
  cutoff.logFC = 1,
  n.labeled.features = 15,
  control.group,
  prefix
)
}
\arguments{
\item{data}{A data frame from RunDEA containing genes and their
corresponding AveExpr, logFC, p.value, adj.p.value ect. Mandatory columns
are "logFC", "adj.P.Val" and "HUGO_ID." HUGO_ID could be obtained using
AddGeneName function.}

\item{cutoff.FDR}{a numeric value for adj.P.Val cutoff. Default = 0.01}

\item{cutoff.logFC}{a numeric value for logFC cutoff. Default = 1}

\item{n.labeled.features}{a number of the most up-/down- regulated features
labeled in the volcano plot. Default = 15}

\item{control.group}{A string vector defining control group name for each
dataset (e.g., "healthy" or "normal").}

\item{prefix}{a prefix for the output filename.}
}
\value{
several .png files with:
1) Volcano plot of differentially expressed features saved into .png
2) Venn Diagram showcasing the size of intersections of the features
between the sample groups
3) Upset plot showcasing the size of intersections of the features
between the sample groups
}
\description{
This function wraps several functions for the visualisation
of DEA results:
1) A function for making volcano plots from DEA results.
The input is typically an output from RunDEA (re-formatted matrix of
differential expression/abundance results from limma, please, check RunDEA
for more details). By default, top 15 up- and down-regulated features
will be labeled on the Volcano plot. Important: Features are selected from
the table with all the groups' comparisons.
2) A function for making Venn diagrams to visualize the size of
intersections of the features between the groups
(e.g., differentially expressed genes for each subtype)
3) A function for making Upset plots to visualize the size of
intersections of the features present each group
(e.g., differentially expressed genes for each subtype)
}
\examples{
{
RunDEAVisuals(campp2_brca_1_DEA_HUGO, cutoff.FDR = 0.05, cutoff.logFC = 1,
n.labeled.features = 15, control.group= "healthy", prefix="test_DEA_visuals")
}
}
