% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_RF.R
\name{RunRF}
\alias{RunRF}
\title{A wrapper function for random forest classification and variable selection}
\usage{
RunRF(
  data,
  group,
  split.size,
  test.train.ratio,
  num.trees.init = 5000,
  num.trees.iterat = 2000
)
}
\arguments{
\item{data}{a matrix of (transformed and normalized) feature counts from
"seq", "array", "ms" or "other" technology (with feature IDs as row names
and sample IDs as columns). If available, batch corrected data should be used.}

\item{group}{a factor specifying group for each sample (e.g. could be
represented by a column from a metadata file)}

\item{split.size}{an integer specifying the minimum number of samples that
the groups must contain in order to carry out random forest classification
and validation}

\item{test.train.ratio}{a floating point number between 0 and 1 representing
the ratio of samples to keep as validation dataset. For example, a
test.train.ratio = 0.25 splits 25 percent of the data into a validation dataset,
meaning 75 percent of the data will be kept as the training dataset.}

\item{num.trees.init}{an integer specifying number of trees to use for the first
forest in the feature selection process. Default is 5000.}

\item{num.trees.iterat}{an integer specifying number of trees to use for
all additional forests in the feature selection process. Default is 2000.}
}
\value{
a list containing two elements:
1) The first element is a list containing the output from the RFApply function
2) The second element is a character vector containing the intersection of
selected variables (e.g., genes) from the feature selection process
}
\description{
The function runs random forest classification and variable
selection by running the RFApply function. The classification and variable
selection are calculated 10 times for 10 random seeds and the intersection of
the selected variables (e.g., genes) is obtained.
The function optionally performs validation using a validation
dataset of the random forest classification if there is enough samples in the
dataset. This is determined by the split.size argument. If validation is not
performed, only variable selection using random forest is carried out.
The function also intersects the selected features across all 10 seed runs
as found during the feature selection process.
}
\examples{
\dontrun{
campp2_brca_1_batchCorrected<-BatchCorrect(data=campp2_brca_1_normalized,
batch=campp2_brca_1_meta$tumor_stage,group=campp2_brca_1_meta$diagnosis,
technology="seq")

campp2_brca_1_run_rf <-
RunRF(data = campp2_brca_1_batchCorrected,
group = campp2_brca_1_meta$diagnosis,
split.size = 5,
test.train.ratio = 0.25,
num.trees.init = 5000,
num.trees.iterat = 2000)
}
}
