% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split_list.R
\name{SplitList}
\alias{SplitList}
\title{Split list}
\usage{
SplitList(my.list)
}
\arguments{
\item{my.list}{list}
}
\value{
split list
}
\description{
Spliting list of files using ","
}
\examples{
\dontrun{
...
}
}
