% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_trim_interactinos.R
\name{TrimWriteInt}
\alias{TrimWriteInt}
\title{Write and trim interactions}
\usage{
TrimWriteInt(my.nodes.list)
}
\arguments{
\item{my.nodes.list}{A list of networks, i.e. output of the function "GetGeneMiRNAInt" or ourput of "GetDeaPPInt" or output of "MatchPPGmiRInt".}
}
\description{
A function to write out interaction lists and trim interactions for plotting
}
\examples{
\dontrun{
...
}

}
