% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_analysis.R
\name{WGCNAAnalysis}
\alias{WGCNAAnalysis}
\title{WGCNA Analysis}
\usage{
WGCNAAnalysis(my.data, my.thresholds, my.name)
}
\arguments{
\item{my.data}{a dataframe with expression/abundance values}

\item{my.thresholds}{a vector of length two. First argument specifies the minimum size of a module and the second argument specifies the dissimilarity cutoff for merging of modules.}

\item{my.name}{a name of output plot(s)}
}
\value{

}
\description{
Weighted gene co-expression network analysis for describing the correlation patterns among genes
}
\examples{
\dontrun{
...
}
}
\seealso{

}
