% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGeneName.R
\name{AddGeneName}
\alias{AddGeneName}
\title{Add Gene Name function}
\usage{
AddGeneName(data, ensembl.version)
}
\arguments{
\item{data}{a data matrix containing a column of stable gene IDs (column must be called 'name').}

\item{ensembl.version}{This arugment specifies which ensembl database to use when transforming ensemble IDs into HUGO IDs using biomaRt. The argument should be specified as a number.}
}
\value{
the input dataframe merged with a column of HUGO IDs.
}
\description{
A function for adding a column of HUGO IDs to a dataframe with a column of Ensemble IDs using biomaRt.
}
\examples{
\dontrun{
...
}
}
