% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGeneName.R
\name{AddGeneName}
\alias{AddGeneName}
\title{Add Gene Name function}
\usage{
AddGeneName(data)
}
\arguments{
\item{data}{a data matrix containing a column of stable gene IDs (column must be called 'name').}
}
\value{
the input data matrix merged with a gene name column.
}
\description{
A function for transforming stable gene IDs into HUGO IDs using biomaRt.
}
\examples{
\dontrun{
...
}
}
