% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_normalization.R
\name{applyNormalization}
\alias{applyNormalization}
\title{Apply Normalization and Transformation}
\usage{
applyNormalization(
  data,
  data.original = NULL,
  group,
  transform,
  standardize,
  technology
)
}
\arguments{
\item{transform}{a string vector of length 1 (or two in case of 2 datasets) defining transformation type for each dataset ("log2", "logit", "log10" or voom).}

\item{standardize}{a string vector of length 1 (or two in case of 2 datasets) defining standardization method  ("mean" or "median"); By default, data from "seq" technology are normalized by "TMM"; "array" technology is standardized using "quantile"}

\item{technology}{a string vector of length 1 (or two in case of 2 datasets) defining technology used for generating the data. Allowed types are: 'array', 'seq', 'ms' or 'other'.}

\item{data1}{a dataframe with expression/abundance counts (genes as rows; samples as columns)}

\item{group1}{a dataframe with groups metadata for each sample (given from metadata table)}

\item{group2}{a dataframe with groups metadata for each sample from a second dataset}

\item{data1.original}{a original data1 including zero values}

\item{data2}{a dataframe with expression/abundance counts for a second dataset}

\item{data2.original}{a original data2 including zero values}
}
\value{
Elist of normalized, filtered and transformed (gene) counts data
}
\description{
A function for applying Normalization ("mean" or "median") and Transformation ("log2", "logit", "log10" or voom). This pipeline does not handle background correction of single-channel intensity data1 or within array normalization two-color intensity data. See limma manual section on array normalization for more on this. Data may be fully normalized with limma or another software and the pipeline re-run.
}
\examples{
\dontrun{
...
}

}
\seealso{

}
