% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_correction.R
\name{BatchCorrect}
\alias{BatchCorrect}
\title{Batch correction}
\usage{
BatchCorrect(data, batch, group, technology)
}
\arguments{
\item{data}{Elist (seq technology) or a matrix (array, ms, other technologies) of normalized, filtered and transformed (gene) counts data}

\item{batch}{a factor derived from metadata column including information about a batch for each sample from data}

\item{group}{a factor derived from metadata column selected as a sample group (e.g. diagnosis)}

\item{technology}{a string vector of length 1 defining technology used for generating the data. Allowed types are: 'array', 'seq', 'ms' or 'other'.}
}
\value{
a list including matrix (array) of batch corrected feature counts
}
\description{
Function for a batch correction. Batch corrected data are NOT intended for a DEA analysis with limma.
}
\examples{
\dontrun{
...
}
}
