% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_correction.R
\name{batchCorrect}
\alias{batchCorrect}
\title{Batch correction}
\usage{
batchCorrect(
  data,
  sdata = NULL,
  batch = NULL,
  sbatch = NULL,
  databatch = NULL,
  sdatabatch = NULL,
  group,
  sgroup = NULL,
  technology
)
}
\arguments{
\item{data}{a gene counts matrix}

\item{sdata}{a gene counts matrix for a secend dataset}

\item{batch}{batch information for each sample}

\item{sbatch}{batch information for each sample from the second dataset}

\item{databatch}{TRUE/FALSE parameter for batch correction}

\item{sdatabatch}{TRUE/FALSE parameter for batch correction for the second dataset}

\item{group}{group information for each sample}

\item{sgroup}{group information for each sample from the second dataset}

\item{technology}{technology}
}
\value{
batch corrected gene counts
}
\description{
Function for batch correction
}
\examples{
\dontrun{
...
}
}
\seealso{

}
