% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_correction.R
\name{BatchCorrect}
\alias{BatchCorrect}
\title{Batch correction}
\usage{
BatchCorrect(data, batch, group, technology)
}
\arguments{
\item{data}{an Elist ("seq" technology) or a matrix ("array", "ms", "other" technologies) of normalized and transformed feature (gene) counts data}

\item{batch}{a factor specifying batch for each sample (e.g. could be represented by a column from a metadata file)}

\item{group}{a factor specifying group for each sample (e.g. could be represented by a column from a metadata file)}

\item{technology}{a string vector of length 1 defining technology used for generating the data. Allowed types are: "array", "seq", "ms" or "other"}
}
\value{
a list including matrix (array) of batch corrected feature counts
}
\description{
This function removes batch effects (using ComBat function) which could be causing a significant heterogeneity across batches of data. Batch corrected data are intended for explanatory purposes, NOT for a DEA analysis.
}
\examples{
{
campp2_brca_1_batchCorrected<-BatchCorrect(data=campp2_brca_1_normalized,
batch=campp2_brca_1_meta$tumor_stage,group=campp2_brca_1_meta$diagnosis,
technology="seq")
}
}
