% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_batchCorrected_mean}
\alias{campp2_brca_1_batchCorrected_mean}
\title{campp2_brca_1_batchCorrected_mean - example data}
\format{
a list of 2 data frames (both 9048 x 7)
}
\usage{
data(campp2_brca_1_batchCorrected_mean)
}
\value{
a list of 2 data frames (both 9048 x 7)
}
\description{
This object includes information about an average counts and SD for
each feature across all the samples and for each sample group. Results are
provided in a form of a list of data frames.
Data were generated by running:
campp2_brca_1_batchCorrected_mean<-
meanCounts(campp2_brca_1_batchCorrected, campp2_brca_1_meta$diagnosis)
}
\keyword{datasets}
