% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_DEA}
\alias{campp2_brca_1_DEA}
\title{campp2_brca_1_DEA - example data}
\format{
a list of:
1) a data frame (43 x 8) - re-formatted matrix of differential
expression/abundance results from limma.
2) a list of original results (a list) from limma
3) a character vector with unique feature names
4) a design matrix (8x4)
5) a contrast matrix (4x1)
}
\usage{
data(campp2_brca_1_DEA)
}
\value{
a list of:
1) a data frame (43 x 8) - re-formatted matrix of differential
expression/abundance results from limma.
2) a list of original results (a list) from limma
3) a character vector with unique feature names
4) a design matrix (8x4)
5) a contrast matrix (4x1)
}
\description{
Results from differential gene expression analysis provided by
RunDEA function using the provided example script:
campp2_brca_1_DEA<-RunDEA(data=campp2_brca_1, metadata=campp2_brca_1_meta,
group=campp2_brca_1_meta$subtype, prefix="test",
block=NULL, batch=campp2_brca_1_meta$age,
covarDEA = c("tumor_stage"), cutoff.logFC=1, cutoff.FDR=0.05)
}
\keyword{datasets}
