% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_forest_features}
\alias{campp2_brca_1_forest_features}
\title{campp2_brca_1_forest_features.rda - example data}
\format{
a list containing four elements
\describe{
}
}
\usage{
campp2_brca_1_forest_features
}
\description{
This object represents output from the
Forest_Features function which is used to fit random forest model
to the data and to perform feature selection using random forest.
Data were generated by running:
campp2_brca_1_forest_features <- Forest_Features(seed = 173,
data = campp2_brca_1,
group = campp2_brca_1_meta$diagnosis,
validation = TRUE,
num_trees = 30,
num_trees_iterat = 15)
}
\keyword{datasets}
