% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_forest_features}
\alias{campp2_brca_1_forest_features}
\title{campp2_brca_1_forest_features.rda - example data}
\format{
a list containing four elements
}
\usage{
data(campp2_brca_1_forest_features)
}
\value{
a list containing four elements
}
\description{
This object represents output from the
ForestFeatures function which is used to fit random forest model
to the data and to perform feature selection using random forest.
Data were generated by running:
campp2_brca_1_forest_features <- ForestFeatures(seed = 123,
data = campp2_brca_1_batchCorrected,
group = campp2_brca_1_meta$diagnosis,
validation = TRUE,
test.train.ratio = 0.25,
num.trees.init = 5000,
num.trees.iterat = 2000)
}
\keyword{datasets}
