% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_LASSO}
\alias{campp2_brca_1_LASSO}
\title{campp2_brca_1_LASSO - example data}
\format{
a list of 3 data frame:
1) (42x1) and 2 empty variables
2) NA
3) NA
}
\usage{
data(campp2_brca_1_LASSO)
}
\value{
#' a list of 3 data frame:
1) (42x1) and 2 empty variables
2) NA
3) NA
}
\description{
This object represents an output from runLASSO function. The list
includes a dataframe of feature names detected by both DEA and EN/LASSO/Ridge
regression. Another 2 sub-objects from the list for cross validation error
and roc. are empty because the input data
is not large enough for the computation of those metrics.
Data were generated by running because LASSO can be run on data where there
is at least 8 samples per group. Here we create a large dataset:
campp2_test_data_LASSO<-cbind(campp2_brca_1,campp2_brca_2)
campp2_test_metadata_LASSO<-rbind(campp2_brca_1_meta, campp2_brca_2_meta)
campp2_test_data_LASSO_replaceNAs<-ReplaceNAs(data=campp2_test_data_LASSO)
campp2_test_data_LASSO_zeroFix<-FixZeros(data=campp2_test_data_LASSO_replaceNAs,group=campp2_test_metadata_LASSO$diagnosis, remove.sparse.features=TRUE)
campp2_test_data_LASSO_normalized<-NormalizeData(data=campp2_test_data_LASSO_zeroFix,group=campp2_test_metadata_LASSO$diagnosis,standardize="TMM",transform="voom",technology="seq")
campp2_test_data_LASSO_batchCorrected<-BatchCorrect(data=campp2_test_data_LASSO_normalized,batch=campp2_test_metadata_LASSO$tumor_stage,group=campp2_test_metadata_LASSO$diagnosis,technology="seq")
run lasso on a large dataset
campp2_brca_1_LASSO<-runLASSO(data=campp2_test_data_LASSO_batchCorrected, group=campp2_test_metadata_LASSO$diagnosis, alpha=0.5, min.coef=0, nfolds=10, prefix="test")
}
\keyword{datasets}
