% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_meta}
\alias{campp2_brca_1_meta}
\title{campp2_brca_1_meta - example data}
\format{
A data frame with 25 rows (samples) and 9 columns.
\describe{
}
}
\usage{
campp2_brca_1_meta
}
\description{
A metadata for campp2_brca_1 containing 8 variables for 25 samples (mix of tumor x normal samples). Data are derived from TCGA dataset.
}
\keyword{datasets}
