% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_1_rf_apply}
\alias{campp2_brca_1_rf_apply}
\title{campp2_brca_1_rf_apply.rda - example data}
\format{
a list containing eight elements
}
\usage{
data(campp2_brca_1_rf_apply)
}
\value{
a list containing eight elements
}
\description{
This object represents output from the
RFApply function which is used to run the ForestFeatures function
across seeds.
Data were generated by running:
campp2_brca_1_rf_apply <-
RFApply(data = campp2_brca_1_batchCorrected,
group = campp2_brca_1_meta$diagnosis,
validation = TRUE,
test.train.ratio = 0.25,
num.trees.init = 5000,
num.trees.iterat = 2000)
}
\keyword{datasets}
