% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_2}
\alias{campp2_brca_2}
\title{campp2_brca_2 - example data}
\format{
A data frame with 10000 rows (genes) and 8 columns (samples)
}
\usage{
data(campp2_brca_2)
}
\value{
A data frame with 10000 rows (genes) and 8 columns (samples)
}
\description{
A campp2_brca_2 contains 10000 gene counts for 8 samples: 4 tumor (1
subtype) x 4 normal samples). Data are derived
from TCGA dataset.
}
\keyword{datasets}
