% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_DE_feature.R
\name{DAFeature}
\alias{DAFeature}
\title{Differentially abundant features - DA analysis}
\usage{
DAFeature(my.contrast, my.data, my.design, coLFC, coFDR, my.block = NULL)
}
\arguments{
\item{my.contrast}{a contrast between groups of interest}

\item{my.data}{a dataframe of expression/abundance counts}

\item{my.design}{a design matrix with all comparisons}

\item{my.block}{if blocking than a vector of patient IDs}

\item{my.coLFC}{a cutoff for logFC}

\item{my.coFDR}{a cutoff for FDR}

\item{my.vector}{a vector of patient IDs; TRUE/FALSE statement specifying output format, if TRUE the function return a vector of feature IDs only}
}
\value{
DA/DE features for DAFeatureApply function
}
\description{
Function for returning DA/DE features for DAFeatureApply function
}
\examples{
\dontrun{
...
}
}
