% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_Feature_apply.R
\name{DAFeatureApply}
\alias{DAFeatureApply}
\title{Apply DAFeature}
\usage{
DAFeatureApply(
  my.contrasts,
  my.data,
  my.design,
  coLFC,
  coFDR,
  my.block = NULL,
  my.vector
)
}
\arguments{
\item{my.contrasts}{= a contrast between groups of interest}

\item{my.data}{= a dataframe of expression/abundance counts}

\item{my.design}{= a design matrix with all comparisons}

\item{my.block}{if blocking than a vector of patient IDs}

\item{my.vector}{a vector of patient IDs; TRUE/FALSE statment specifying output format, if TRUE the function return a vector of feature IDs only}

\item{my.coLFC}{a cutoff for logFC}

\item{my.coFDR}{a cutoff for FDR}
}
\value{
DA/DE features
}
\description{
Apply differential abundance analysis to all the comparisons at once
}
\examples{
\dontrun{
DAFeatureApply(contrast.matrix, arg.data, arg.design, arg.logFC, arg.FDR, NULL, FALSE)
}
}
\seealso{

}
