% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_STRING.R
\name{DownloadPPInt}
\alias{DownloadPPInt}
\title{Download string database}
\usage{
DownloadPPInt(my.geneIDs, my.version = "11.0")
}
\arguments{
\item{my.geneIDs}{a string specifying type of gene ID to use to match IDs in STRING database. Options are: ensembl_peptide_id, hgnc_symbol, ensembl_gene_id, ensembl_transcript_id or uniprotswissprot.}

\item{my.version}{Version of STRING database to use. Default is version 11.}
}
\description{
Function to download string database
}
\examples{
\dontrun{
...
}
}
