% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_K_means.R
\name{EstimateKmeans}
\alias{EstimateKmeans}
\title{Estimate Kmeans}
\usage{
EstimateKmeans(data)
}
\arguments{
\item{data}{a dataframe of feature (e.g. gene) counts}
}
\value{
number of clusters
}
\description{
A function for estimating K-means.
Number of kmeans will be based on the bayesian information criterion(BIC)
provided by mclust package. Summary describing the best model is printed on
the screen during the calculation.
}
\examples{
\dontrun{
...
}
}
