% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_K_means.R
\name{EstimateKmeans}
\alias{EstimateKmeans}
\title{Estimate Kmeans}
\usage{
EstimateKmeans(df, n)
}
\arguments{
\item{n}{= number of sample subsets to generate (The number of clusters tested will be based on number of samples, fewer samples will result in fewer kmeans tested.)}

\item{my.data}{a dataframe of expression/abundance counts}
}
\value{
clustering results and plots
}
\description{
Function for Estimating Kmeans; Number of kmeans will be based on the bayesian information criterion(BIC); RNAseq with
many genes, multiple samples of 3000 variables will be generated
and tested to overcome issues with computational time and the
consensus of best n kmeans will be returned.
Clustering may only be performed one dataset at a time!
}
\examples{
\dontrun{
...
}
}
\seealso{

}
