% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_K_means.R
\name{EstimateKmeans}
\alias{EstimateKmeans}
\title{Estimate K-means clusters}
\usage{
EstimateKmeans(data)
}
\arguments{
\item{data}{a data frame of feature (e.g. gene) counts}
}
\value{
a data frame with a number of clusters
}
\description{
A function to estimate the number of K-means.
Number of K-means is based on the Bayesian information criterion (BIC)
provided by mclust package. A summary describing the best model is printed on
the screen during the calculation.
}
\examples{
\dontrun{
EstimateKmeans(t(campp2_brca_1_batchCorrected[1:2000,]))
}
}
