% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_DEA.R
\name{ExportDEA}
\alias{ExportDEA}
\title{Reformat and Export DEA results into .txt}
\usage{
ExportDEA(res.DEA, prefix)
}
\arguments{
\item{res.DEA}{a list of data frames containing results from different group
comparisons in limma format.}

\item{prefix}{a character vector defining prefix of output file name.}
}
\value{
1) a data frame of DEA results
2) a .txt table of exported DEA results
}
\description{
A function for re-formatting the output from DEA generated based
on the workflow: (runDEA - DEAFeatureApply - DEAFeature - limma); and
exporting the results into a .txt file.
}
\examples{
{
campp2_brca_1_DEA_out<-ExportDEA(res.DEA = campp2_brca_1_DEA$res.DEA,
prefix="test")
}
}
