% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting_data_distribution.R
\name{FitDistributions}
\alias{FitDistributions}
\title{Fitting Data Distributions}
\usage{
FitDistributions(my.data)
}
\arguments{
\item{my.data}{a dataframe of expression/abundance counts, N.B only a subset of variables should be input, not intended for the full expression matrix!}
}
\value{
data for distribution plots
}
\description{
A function for fitting data distribution
}
\examples{
\dontrun{
...
}
}
