% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting_data_distribution.R
\name{FitDistributions}
\alias{FitDistributions}
\title{Fitting Data Distributions}
\usage{
FitDistributions(data)
}
\arguments{
\item{data}{a data frame of feature counts. Only a subset of
features should be input, not intended for the full feature count matrix!}
}
\value{
a list of the results from fitdist function describing distribution
of the data
}
\description{
A function for fitting data distribution.
Depending on character of the data, various distributions are calculated.
In case of only integer numbers present in the dataset, poison and normal
distributions are tested. In case of presence of float values in the dataset,
Weibull, gamma, log normal and normal distributions are tested. In case of
negative float values, only normal distribution is calculated.
}
\examples{
{
campp2_brca_1_distributionsFit <- FitDistributions(campp2_brca_1[1:10,])
}
}
