% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASSO.R
\name{LASSOFeature}
\alias{LASSOFeature}
\title{A function for LASSO/Elastic net/Ridge regression}
\usage{
LASSOFeature(
  seed = 123,
  data,
  group,
  alpha = 0.5,
  validation = FALSE,
  min.coef = 0,
  nfolds = 10
)
}
\arguments{
\item{seed}{an integer vector containing the random number generator (RNG)
state. Default is 123.}

\item{data}{a data frame of expression/abundance counts.
It's recommended to use normalized and batch corrected data.}

\item{group}{a factor specifying group for each sample (e.g. could be
represented by a column from a metadata file)}

\item{alpha}{a numeric vector specifying hyperparameter alpha. This value
must be set to 0.0 < x < 1.0 for Elastic Net (0.5 is default) or to 1.0 for
LASSO regression or to 0.0 for Ridge regression.}

\item{validation}{a boolean value for performing double cross-validation.
Validation (test) set represents 1/4 of the samples in each group.
As a result, cross validation error rate is provided. Default value is FALSE.}

\item{min.coef}{a numeric vector specifying a threshold for features'
filtering (e.g. genes) based on the coefficients which are calculated during
model fitting. Default value is > 0.}

\item{nfolds}{a numeric vector describing number of folds during Lambda
estimation which is based on a cross-validation. Although nfolds can be
as large as the sample size (leave-one-out CV), it is not recommended for
large datasets. Smallest value allowable is nfolds=3. Default is 10.}
}
\value{
a list of:
1) coef.ma.names - a matrix of feature names having coefficients of best model
passing the filters (threshold defined by min.coef)
2) class.error - logical/numeric value describing miss classification error
rate
3) fit - cv.glmnet object
4) coef.ma - a matrix of the features' coefficients (best model)
passing the filters (threshold defined by min.coef)
}
\description{
The function runs LASSO/Elastic net/Ridge regression
(depending on a hyperparameter alpha value). Parameter Lambda is
automatically estimated using k-fold cross-validation (cv.glmnet) by default.
The function optionally calculates also double cross-validation using
a test set. aValidation (test) set represents 1/4 of the samples from each
group.
Results are provided in the form of:
1) the ames of the signifficant features passing filters based on their
coefficient values
2) classification error rate (calculated if "validation=TRUE")
3) cv.glmnet object.
}
\examples{
{
LASSOFeature(seed=123, data=campp2_brca_1,
group=campp2_brca_1_meta$diagnosis, alpha=0.5, validation=TRUE,
min.coef = 0, nfolds=10)
}
}
