% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{MakeHeatmap}
\alias{MakeHeatmap}
\title{Heatmaps}
\usage{
MakeHeatmap(
  my.DE,
  my.gradient,
  my.colorshm,
  my.colors,
  my.group,
  my.filename,
  my.range
)
}
\arguments{
\item{my.DE}{a dataframe with counts for differential expressed/abundant features}

\item{my.gradient}{a color gradient to use for heatmap}

\item{my.colorshm}{a color pallet for groups full length}

\item{my.colors}{a color pallet for groups at levels}

\item{my.group}{a vector of groups to color by}

\item{my.filename}{a name of output plot}

\item{my.range}{??}
}
\value{
heatmaps
}
\description{
Function for making heatmaps
}
\examples{
\dontrun{
...
}
}
