% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{MakeHeatmap}
\alias{MakeHeatmap}
\title{Make Heatmap}
\usage{
MakeHeatmap(data, gradient, group, prefix, range)
}
\arguments{
\item{data}{A dataframe with counts for differential expressed/abundant features.}

\item{gradient}{A color gradient to use for the heatmap.}

\item{group}{A vector of groups to color by.}

\item{prefix}{A prefix for the output filename.}

\item{range}{A vector containing the smallest and biggest logFC values from the input data frame.}
}
\value{
heatmap
}
\description{
A function for making heatmaps from expression data.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
