% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpsetPlot.R
\name{MakeUpset}
\alias{MakeUpset}
\title{Upset Plot}
\usage{
MakeUpset(prefix, sets_list, names_sets)
}
\arguments{
\item{prefix}{A prefix for the output filename.}

\item{sets_list}{A list of sets to compare.}

\item{names_sets}{A vector of the corresponding names to each set in sets_list.}
}
\value{
Upset Plot
}
\description{
A function for making Upset plots to visualize the intersection between 2 or more sets.
}
\examples{
\dontrun{
...
}
}
