% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpsetPlot.R
\name{MakeUpset}
\alias{MakeUpset}
\title{Upset Plot}
\usage{
MakeUpset(
  prefix,
  groups.feature.list,
  label = "",
  label.vertical.position = 15,
  y.axis.by = 100,
  set.size.by = 200
)
}
\arguments{
\item{prefix}{a prefix for the output filename.}

\item{groups.feature.list}{a list of character vectors (each character
vector named by its group name, e.g., "LumA") including feature names
characteristic for each group of interest (e.g., subtype)}

\item{label}{a character vector used as a label for the upset plot. Default
is "".}

\item{label.vertical.position}{a numeric value describing vertical position of
the label. Default is 15.0 (cm)}

\item{y.axis.by}{an integer describing interval used for labeling Y-axis.
Defaults is 100.}

\item{set.size.by}{an integer describing interval used for labeling the
set size axis. Defaults is 200.}
}
\value{
an Upset plot showcasing the size of intersections of the features
between the sample groups
}
\description{
A function for making Upset plots to visualize the size of
intersections of the features present each group
(e.g., differentially expressed genes for each subtype).
}
\examples{
{
control.group="healthy"
groups.full.list <- split.data.frame(campp2_brca_1_DEA_HUGO, campp2_brca_1_DEA_HUGO$comparison)
campp2_brca_1_DEA_HUGO_features_per_group=list()
for (features in groups.full.list){
    campp2_brca_1_DEA_HUGO_features_per_group <- append(campp2_brca_1_DEA_HUGO_features_per_group,list(features$name))
}
names(campp2_brca_1_DEA_HUGO_features_per_group) <- gsub("-","",gsub(control.group,"",names(groups.full.list)))
MakeUpset("testUpSet",campp2_brca_1_DEA_HUGO_features_per_group,
label="test_UpSet",label.vertical.position=16.5,
y.axis.by=300,set.size.by=250)
}
}
