% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennDiagram.R
\name{MakeVennDiagram}
\alias{MakeVennDiagram}
\title{Venn Diagram}
\usage{
MakeVennDiagram(prefix, sets_list, names_sets)
}
\arguments{
\item{prefix}{A prefix for the output filename.}

\item{sets_list}{A list of sets containing the gene names for each subtype to compare.}

\item{names_sets}{A vector of the corresponding names of each subtype in sets_list. names_sets must be same length as sets_list.}
}
\value{
Venn Diagram
}
\description{
A function for making Venn diagrams to visualize the intersection size of significantly differentially expressed genes between 2 or more subtypes. The genes to compare is gathered from the results provided by DEA using CAMPP2.
}
\examples{
\dontrun{
...
}
}
