% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennDiagram.R
\name{MakeVennDiagram}
\alias{MakeVennDiagram}
\title{Venn Diagram}
\usage{
MakeVennDiagram(prefix, groups.feature.list)
}
\arguments{
\item{prefix}{A prefix for the output filename.}

\item{groups.feature.list}{A list of character vectors (each character
vector named by its group name, e.g., "LumA") including feature names
characteristic for each group of interest (e.g., subtype).}
}
\value{
Venn Diagram
}
\description{
A function for making Venn diagrams to visualize the size of
intersections of the features between the groups
(e.g., differentially expressed genes for each subtype)
}
\examples{
{
MakeVennDiagram("test_Venn", campp2_brca_1_DEA_HUGO_features_per_group)
}
}
