% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennDiagram.R
\name{MakeVennDiagram}
\alias{MakeVennDiagram}
\title{Venn Diagram}
\usage{
MakeVennDiagram(prefix, groups.feature.list)
}
\arguments{
\item{prefix}{A prefix for the output filename.}

\item{groups.feature.list}{A list of character vectors (each character
vector named by its group name, e.g., "LumA") including feature names
characteristic for each group of interest (e.g., subtype).}
}
\value{
Venn Diagram
}
\description{
A function for making Venn diagrams to visualize the size of
intersections of the features between the groups
(e.g., differentially expressed genes for each subtype)
}
\examples{
{
control.group="healthy"
groups.full.list <- split.data.frame(campp2_brca_1_DEA_HUGO, campp2_brca_1_DEA_HUGO$comparison)
campp2_brca_1_DEA_HUGO_features_per_group=list()
for (features in groups.full.list){
    campp2_brca_1_DEA_HUGO_features_per_group <- append(campp2_brca_1_DEA_HUGO_features_per_group,list(features$name))
}
names(campp2_brca_1_DEA_HUGO_features_per_group) <- gsub("-","",gsub(control.group,"",names(groups.full.list)))

MakeVennDiagram("test_Venn", campp2_brca_1_DEA_HUGO_features_per_group)
}
}
