% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{MakeVolcano}
\alias{MakeVolcano}
\title{Make Volcano Plot}
\usage{
MakeVolcano(data, prefix)
}
\arguments{
\item{data}{A data matrix from RunDEA containing genes, AveExpr, logFCs,p.value, adj.p.value ect.}
}
\value{
Volcano plot
}
\description{
A function for making volcano plots from DEA results.
}
\examples{
\dontrun{
...
}
}
