% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{MakeVolcano}
\alias{MakeVolcano}
\title{Make Volcano Plot}
\usage{
MakeVolcano(data, prefix, cutoff.logFC, cutoff.FDR)
}
\arguments{
\item{data}{A data matrix from RunDEA containing genes and their corresponding AveExpr, logFCs, p.value, adj.p.value ect. Mandatory columns are logFC, adj.P.Val and HUGO_ID. N.B. columns logFC and adj.P.Val must contain numeric values!}

\item{prefix}{A prefix for the output filename.}

\item{cutoff.logFC}{A cut-off value for log fold change}

\item{cutoff.FDR}{A cut-off value for  corrected adj.P.Val}
}
\value{
Volcano plot
}
\description{
A function for making volcano plots from DEA results.
}
\examples{
\dontrun{
...
}
}
