% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_data.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalization and Transformation}
\usage{
NormalizeData(data, group, transform, standardize, technology)
}
\arguments{
\item{data}{a dataframe of expression/abundance counts}

\item{group}{a factor specifying sample group from metadata}

\item{transform}{a string vector of length 1 defining transformation type for each dataset ("log2", "logit", "log10"). If technology is "seq", this option is ignored an Voom transform is applied instead.}

\item{standardize}{a string vector of length 1 defining standardization method ("mean" or "median"). If technology is "seq" or "array", this option is ignored, instead, seq data is normalized by "TMM"; array data is normalized by "quantile" standardization.}

\item{technology}{a string vector of length 1 defining technology used for generating the data. Allowed types are: 'array', 'seq', 'ms' or 'other'.}
}
\value{
Elist (seq technology) or array (array, ms, other technologies) of normalized, filtered and transformed (gene) counts data
}
\description{
This function normalizes and transforms feature counts depending on data types and selected normalization/standardization methods.
}
\examples{
\dontrun{
...
}

}
