% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{PCAPlot}
\alias{PCAPlot}
\title{Principal Component Analysis (PCA)}
\usage{
PCAPlot(data, group, PCA.labels = "none", cols = NULL, prefix = "")
}
\arguments{
\item{data}{a dataframe of expression/abundance counts}

\item{group}{a factor specifying group for each sample (e.g. could be
represented by a column from a metadata file)}

\item{PCA.labels}{a text ("all", "none) specifying the elements to be
labelled. Default value is "none"}

\item{cols}{a vector of colors (one color for each group)}

\item{prefix}{a character defining a prefix of output file}
}
\value{
1) scree plot, 2) contributions of variables to PC1, 3) contributions
of variables to PC2, 4) 2D PCA plot, 5) 3D PCA plot
}
\description{
A function calculating PCA. PCA plot shows the relationships
between samples (based on squared euclidean distances) in the data set.
Each dot represents one sample. This function provides also a scree plot
describing percentages of explained variances by each PCA component and
plots with contributions of variables to PCA1 and PCA2. Top 3 PCA components
are visualized in 3D plot.
}
\examples{
\dontrun{
PCAPlot(campp2_brca_1_batchCorrected, as.factor(campp2_brca_1_meta$subtype), PCA.labels ="none", cols=NULL, prefix="test_PCA")
}
}
