% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_distributions.R
\name{PlotDistributions}
\alias{PlotDistributions}
\title{Plotting Distributions}
\usage{
PlotDistributions(my.data, list.of.lists)
}
\arguments{
\item{my.data}{a dataframe of expression/abundance counts, N.B only a subset of variables should be input, not intended for the full expression matrix!(See "Fitting Data Distributions" above)}

\item{list.of.lists}{an output from the function "Fitting Data Distributions". The my.data and list.of.list should have the same dimentions, e.g. length of list == nrows of dataframe.}
}
\value{
distribution plots
}
\description{
Plotting the counts' Distributions
}
\examples{
\dontrun{
...
}
}
