% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_distributions.R
\name{PlotDistributions}
\alias{PlotDistributions}
\title{Plotting Distributions}
\usage{
PlotDistributions(data, fitted.data)
}
\arguments{
\item{data}{a dataframe of expression/abundance counts, N.B only a subset of variables should be input, not intended for the full expression matrix!(See "Fitting Data Distributions" above)}

\item{fitted.data}{a list of lists obtained as output from the function "Fitting Data Distributions". The data and fitted.data should have the same dimensions, e.g. length of list == nrows of dataframe.}
}
\value{
distribution plots
}
\description{
Plotting the counts' Distributions
}
\examples{
\dontrun{
...
}
}
