% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_interaction_network.R
\name{PlotInt}
\alias{PlotInt}
\title{Plot interaction network}
\usage{
PlotInt(my.trimmed.list)
}
\arguments{
\item{my.trimmed.list}{A list of trimmed networks, i.e. output of the function "TrimWriteInt".}
}
\description{
A function plotting interaction networks
}
\examples{
\dontrun{
...
}
}
