% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_files.R
\name{ReadMyFile}
\alias{ReadMyFile}
\title{Read files}
\usage{
ReadMyFile(my.data, data.type)
}
\arguments{
\item{my.data}{a dataframe of expression/abundance counts.}

\item{my.expr}{This param is set to TRUE or FALSE by CAMPP2}
}
\description{
Reading input files
}
\examples{
\dontrun{
...
}
}
