% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_NAs.R
\name{ReplaceNAs}
\alias{ReplaceNAs}
\title{Replace "NA" values}
\usage{
ReplaceNAs(data, pct.NA.row = 70, pct.NA.column = 80)
}
\arguments{
\item{data}{a dataframe of gene/abundance counts.}

\item{pct.NA.row}{a number defining maximal percentage of NA values per row (feature). Rows with a higher percentage of NA values will be removed.}

\item{pct.NA.column}{a number defining maximal percentage of NA values per column (sample). Columns with a higher percentage of NA values will be removed.}
}
\value{
Data frame with replaced "NA" values
}
\description{
Replacing "NA" values in the data (counts) using llsImpute (1st round) and impute.knn (if 1st round wasn't successful or resulted in negative values). Rows with more than 50% missing values will be imputed using the overall mean per sample.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
