% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_NAs.R
\name{ReplaceNAs}
\alias{ReplaceNAs}
\title{Replace "NA" values}
\usage{
ReplaceNAs(data)
}
\arguments{
\item{data}{a dataframe of gene/abundance counts.}
}
\value{
Data frame with replaced "NA" values
}
\description{
Replacing "NA" values in the data (counts).
}
\examples{
\dontrun{
...
}
}
\seealso{

}
