% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_NAs.R
\name{ReplaceNAs}
\alias{ReplaceNAs}
\title{ReplaceNAs}
\usage{
ReplaceNAs(my.data)
}
\arguments{
\item{my.data}{a dataframe of expression/abundance counts.}
}
\value{
Dataframe with replaced NA-values
}
\description{
Checking Data for NA Values.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
