% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_NAs.R
\name{ReplaceNAs}
\alias{ReplaceNAs}
\title{Replace "NA" values}
\usage{
ReplaceNAs(data, pct.NA.row = 70, pct.NA.column = 80)
}
\arguments{
\item{data}{a dataframe of gene/abundance counts.}

\item{pct.NA.row}{a number defining maximal percentage of NA values present in
a feature (e.g. a percentage of the samples having "NA" in a given gene).
Features (rows) with a higher percentage of NA values will be removed (70 by
default).}

\item{pct.NA.column}{a number defining maximal percentage of NA values
present in the sample (e.g. a percentage of the genes having NA in one
sample). Samples (columns) with a higher percentage of NA values will be
removed (80 by default).}
}
\value{
Data frame with replaced "NA" values
}
\description{
In the first step, percentages of NAs in rows (features) and columns (samples)
are calculated. Next, features and samples with high percentage of NAs are
removed (by default, features having >70 % and samples having > 80% of
NAs are removed). Remaining NA values are replaced by imputed values
generated by “local least squares (llsImpute)“. If the first round of missing
values imputation failed or resulted in negative values, an additional
missing value imputation using impute.knn will be performed (rows with more
than 50 percent of missing values will be imputed using the overall mean per sample).
}
\examples{
\dontrun{
n<-nrow(campp2_brca_1)
Na_dataset1<-apply (campp2_brca_1, 2, function(x) {x[sample( c(1:n), floor(n/10))] <- NA; x} )
ReplaceNAs(data=Na_dataset1, pct.NA.row=70, pct.NA.column=80)
}
}
\seealso{

}
