% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_NAs.R
\name{ReplaceNAs}
\alias{ReplaceNAs}
\title{Replace "NA" values}
\usage{
ReplaceNAs(data, pct.NA.row = 70, pct.NA.column = 80)
}
\arguments{
\item{data}{a dataframe of gene/abundance counts.}

\item{pct.NA.row}{a number defining maximal percentage of NA values present
in a feature (e.g. a percentage of samples having "NA" in a given gene).
Features (rows) with a higher percentage of NA values will be removed (70 by
default).}

\item{pct.NA.column}{a number defining maximal percentage of NA values
present in a sample (e.g. a percentage of genes having NA in one sample).
Samples (columns) with a higher percentage of NA values will be removed
(80 by default).}
}
\value{
Data frame with replaced "NA" values
}
\description{
In the first step, percentages of NAs in rows (features) and
columns (samples) are calculated. Features and samples with high
percentage of NAs are removed (by default, features having more than 70
percent and samples having more than 80 percent of NAs are removed).
Remaining NA values are replaced by imputed values generated by “local least
squares (llsImpute)“. If the first round of missing values imputation failed
or resulted in negative values, an additional missing value imputation using
impute.knn will be performed (rows with more than 50 percent of missing
values will be imputed using the overall mean per sample).
}
\examples{
{
ReplaceNAs(data=campp2_brca_1_NAs, pct.NA.row=70, pct.NA.column=80)
}
}
