% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_zeros.R
\name{ReplaceZero}
\alias{ReplaceZero}
\title{ReplaceZeros}
\usage{
ReplaceZero(my.data, my.group)
}
\arguments{
\item{my.data}{a dataframe of expression/abundance counts.}

\item{my.group}{a vector of integers specifying group; should be represented by a column in metadata file.}
}
\value{
Dataframe with replaced zeros
}
\description{
Checking Data for NA Values.
}
\examples{
\dontrun{
...
}
}
\seealso{

}
