% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_DEA.R
\name{RunDEA}
\alias{RunDEA}
\title{Run differential expression/abundance analysis}
\usage{
RunDEA(
  data,
  metadata,
  technology,
  batch,
  covarDEA,
  group,
  logFC,
  FDR,
  prefix,
  block
)
}
\arguments{
\item{data}{A raw gene count matrix from seq, array, ms or other technology (with gene IDs as row names and sample IDs as columns). It's recommended to import gene counts using function "import_counts".}

\item{metadata}{Samples' metadata table is recommended to be imported using function "import_metadata". Metadata must include exactly the same samples sorted in the same order as in a gene counts matrix (data).}

\item{technology}{a string vector of length 1 defining technology used for generating the data. Allowed types are: 'array', 'seq', 'ms' or 'other'.}

\item{batch}{The batch covariate for each data sample, derived from metadata column.}

\item{covarDEA}{Covariates to include in the analysis If multiple of these, they should be specified as a character vector.}

\item{group}{An array of specific sample groups derived from metadata column (e.g. diagnosis)}

\item{logFC}{The logarithmic Fold Change for each data sample (ratio of changes in expression data)}

\item{FDR}{The false discovery rate for each data sample (the corrected p-value)}

\item{prefix}{a character defining the result folder name and prefix of output file names.}

\item{block}{A vector or factor specifying a blocking variable. The block must be of same length as data and contain 2 or more options. For 2 datasets, the block can be defined as a vector of the two seperate blocks.}
}
\value{
a matrix of differential expression/abundance results (containing information on gene regulation and significance of comparison)
}
\description{
Run differential expression/abundance analysis
}
\examples{
\dontrun{
...
}
}
\seealso{

}
