% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDEA_LASSO_consensus.R
\name{RunDEA_LASSO_consensus}
\alias{RunDEA_LASSO_consensus}
\title{A function for creating consensus between DEA and LASSO}
\usage{
RunDEA_LASSO_consensus(
  DEA.out,
  LASSO.results,
  group,
  viridis.palette = "turbo",
  prefix
)
}
\arguments{
\item{DEA.out}{a data frame from RunDEA containing features and their
corresponding AveExpr, logFC, p.value, adj.p.value ect. Mandatory column
is "name"}

\item{LASSO.results}{an object generated by runLASSO function}

\item{group}{a factor specifying group for each sample (e.g. could be represented by a column from a metadata file)}

\item{viridis.palette}{a character vector specifying color gradient to use for
the heatmap. Default option for viridis color palettes is 'turbo'. For more
options, see viridis vignette.}

\item{prefix}{a character string defining a prefix of output file.}
}
\value{
1) a data frame of features common for both DEA and LASSO which are extracted
from DEAout table
2) a .png with a Venn diagram of features' intersections between DEA and LASSO
}
\description{
This function creates a consensus between the results from
DEA and LASSO/Elastic net/Ridge regression. Specifically, only the features
present in LASSO results are kept in the DEA results table and are sorted
based on DEA results.
Additionally, a Venn diagram visualizing overlaps is generated.
}
\examples{
{
campp2_brca_1_DEA_LASSO_consensus <- RunDEA_LASSO_consensus(
    campp2_brca_1_DEA$DEA.out,
    campp2_brca_1_LASSO, group=campp2_brca_1_meta$diagnosis,
   viridis.palette="turbo", "test")
}
}
