% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHeatmap.R
\name{RunHeatmap}
\alias{RunHeatmap}
\title{RunHeatmap}
\usage{
RunHeatmap(
  feature.counts,
  DEA.out,
  group,
  heatmap.size = 30,
  viridis.palette = "turbo",
  plot.heatmap = "DEA",
  prefix
)
}
\arguments{
\item{DEA.out}{a data frame from RunDEA containing genes and their
corresponding AveExpr, logFC, p.value, adj.p.value ect. Mandatory columns
are "logFC".}

\item{group}{a factor specifying group for each sample (e.g. could be
represented by a column from a metadata file). Group's information
will be used in clustering the samples.}

\item{heatmap.size}{an argument specifying how many genes will be selected
to be plotted on the heatmap if plot.heatmap is TRUE. The input must be
specified as an even number.}

\item{viridis.palette}{A=a character vector specifying color gradient to use for
the heatmap. Default option for viridis color palettes is 'turbo'. For more
options, see viridis vignette.}

\item{plot.heatmap}{an argument defining which data will be used for
the selection of the top x features to be plotted on the heatmap.
Options are:"DEA", "LASSO", "EN" or "Consensus".}

\item{prefix}{a character vector defining prefix of output file name.}
}
\description{
A function for making heatmaps to showcase the difference in
expression/abundance of features across samples and sample groups and to
visualise relations between them.
}
\examples{
\dontrun{

}
}
