% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHeatmap.R
\name{RunHeatmap}
\alias{RunHeatmap}
\title{RunHeatmap}
\usage{
RunHeatmap(
  feature.counts,
  DEA.out.heatmap,
  heatmap.size = 30,
  group,
  viridis.palette = "turbo",
  plot.heatmap = "DEA",
  prefix,
  data.type = "expression/abundance"
)
}
\arguments{
\item{feature.counts}{a feature count matrix (ideally normalized and batch
corrected) generated by "seq", "array", "ms" or "other" technology (with
feature IDs as row names and sample IDs as columns). It's recommended to
import feature counts using function "import_counts".}

\item{DEA.out.heatmap}{a sorted data frame representing output from one of:
1) runDEA which is based on the output format from limma.
Data frame sorted e.g. by logFC. Mandatory column is "name".
2) runLASSO (including feature names)
3) consensus of DEA and LASSO/EN/Ridge regression generated by
RunDEA_LASSO_consensus and and including features sorted based on DEA
logFC values). Mandatory column is "name".}

\item{heatmap.size}{an argument specifying how many genes will be selected
to be plotted on the heatmap if plot.heatmap is TRUE. The input must be
specified as an even number. Default is 30.

Please, consider that some of the TOP DEA features might be present multiple
times in the DEA output due to the multiple groups comparisons.
On the heatmap, each feature will be projected only once so the final number
of the projected features migh by lower then heatmap.size.}

\item{group}{a factor specifying group for each sample (e.g. could be
represented by a column from a metadata file). Group's information
will be used in clustering the samples.}

\item{viridis.palette}{a character vector specifying color gradient to use for
the heatmap. Default option for viridis color palettes is 'turbo'. For more
options, see viridis vignette.}

\item{plot.heatmap}{an argument defining which data will be used for
the selection of the top x features to be plotted on the heatmap.
Options are:"DEA", "LASSO", "EN", "Ridge" or "Consensus".}

\item{prefix}{a character vector defining prefix of output file name.}

\item{data.type}{a character vector for labeling units in the legend.
Default is "expression/abundance".}
}
\value{
heatmap
}
\description{
A function for making heatmaps to showcase the difference in
expression/abundance of features across samples and sample groups and to
visualise relations between them.
}
\examples{
{
RunHeatmap(campp2_brca_1_batchCorrected, campp2_brca_1_DEA_HUGO, 40,
campp2_brca_1_meta$subtype, viridis.palette="turbo",
plot.heatmap="DEA", "test_heatmap")
}
}
