% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateKmeans2.R, R/runKmeans.R
\name{runKmeans}
\alias{runKmeans}
\title{Estimate Kmeans}
\usage{
runKmeans(data, PCA.labels = FALSE, cols = NULL, prefix = NULL)

runKmeans(data, PCA.labels = FALSE, cols = NULL, prefix = NULL)
}
\arguments{
\item{data}{a dataframe of feature (e.g. gene) counts}

\item{PCA.labels}{a text ("all", "none) specifying the elements to be
labelled. Default value is "none".}

\item{cols}{a vector of colors (one color for each group)}

\item{prefix}{a character defining a prefix of output file names.}
}
\value{
clustering r

a data frame (metadata table) with cluster info assigned to each sample; Figures saved into png.
}
\description{
A function for estimating K-means.
Number of kmeans will be based on the bayesian information criterion(BIC);

(this is more for runEstimateKmeans)
In case of data sets with many genes,
multiple samples of 3000 variables will be generated
and tested to overcome issues with computational time and the
consensus of best n kmeans will be returned.

Clustering may only be performed one dataset at a time.

(The number of clusters
tested will be based on number of samples, fewer samples will result in fewer
kmeans tested.)

A function for estimating the k-means clusters. The clusters
are visualised on a PCA graph.
In case of data sets with >1000 features,
multiple sets (number of sets = number of features/1000; rounded up to the
integer) will be generated
and tested for the number of clusters estimation using BIC.
Consensus of best n k-means will be returned. Number of randomly selected
features in 1 set
is limited to 2000 (for large sets having more than 2000 features).
In case k-means estimation
using BIC fails, number of clusters will be based on the number of samples
(2-6 clusters for dataset with less than 100 samples; 2-11 for datasets with
101-500 samples and 2-16 clusters for datasets with more than 500 samples).
The clusters visualized on a PCA graph are saved into the png.
}
\examples{
\dontrun{
...
}
\dontrun{
...
}
}
