% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Kmeans.R
\name{runKmeans}
\alias{runKmeans}
\title{K-means clustering}
\usage{
runKmeans(
  data,
  show.PCA.labels = FALSE,
  prefix,
  colors = NULL,
  num.km.clusters = NULL,
  seed = NULL,
  pca.scale = FALSE
)
}
\arguments{
\item{data}{a data.frame of feature (e.g. gene) counts}

\item{show.PCA.labels}{a boolean value (TRUE or FALSE) specifying if elements
(e.g. samples) should be labelled in the PCA plot including information about
the clusters. Labeling is based on column names of the input data.
Default value is FALSE.}

\item{prefix}{a character string defining a prefix of output file.}

\item{colors}{a vector of colors (one color for each cluster)}

\item{num.km.clusters}{a vector of numbers of clusters expected in the data.
Clusters can be represented e.g., by sample groups, sub-types, etc.
If multiple values provided, the function will automatically
perform K-means clustering based on each (expected number of clusters) of
them separately. By default, expected numbers of clusters are calculated
automatically based on BIC (mclust package) applied to sub-sampled data sets.
A default value for this parameter is NULL.}

\item{seed}{a seed for K-means clustering. Defaults is NULL.}

\item{pca.scale}{a boolean, if TRUE then data are scaled to unit variance
during PCA. Default is FALSE}
}
\value{
a list including:
1) a data.frame with cluster information assigned to each sample;
2) a PCA objects including data for generating PCA plots with cluster
information (generated by PCA function from FactoMineR package);
2D PCA plot(s) projecting samples over first 2 principal components saved
into png.
plots of BIC values for each sample sub set are saved into .png file
}
\description{
A wrapper function for K-means clustering.
For data sets with >1000 features, multiple subsets (number of sets = number
of features/1000; rounded up to the higher integer) will be generated,
maximum is 10 subsets. A number of randomly selected features (e.g. genes)
in 1 subset is limited to 2000.
A number of clusters for each sub set is automatically estimated using BIC and
a consensus (based on all the subsets) of best n k-means is returned and used
for clustering.
If K-means estimation using BIC fails, a number of clusters will be based on
the number of samples (2-6 clusters for data set with less than 100 samples;
2-11 for data sets with 101-500 samples, and 2-16 clusters for data sets with
more than 500 samples). The number of cluster can also be supplied manually.
As results, an information about the clusters assigned to each sample and
results PCA (FactoMineR) are provided.
The clusters are visualized on a PCA plot and saved into the png.
Plot of BIC values for each of the sample sub-set is saved into .png file.
A summary describing the best model is printed on the screen during
calculation.
}
\examples{
\dontrun{
runKmeans(campp2_brca_1_batchCorrected[1:2000,], show.PCA.labels = FALSE, colors=NULL,
prefix="test", num.km.clusters=NULL, seed=123, pca.scale=FALSE)
}
}
