% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Kmeans.R
\name{runKmeans}
\alias{runKmeans}
\title{K-means clustering}
\usage{
runKmeans(
  data,
  PCA.labels = FALSE,
  cols = NULL,
  prefix = NULL,
  num.km.clusters = NULL
)
}
\arguments{
\item{data}{a data.frame of feature (e.g. gene) counts}

\item{PCA.labels}{a string vector ("all" or "none") specifying which elements
(e.g. samples) should be labelled (labeling is based on column names of the
input data). Default value is "none".}

\item{cols}{a vector of colors (one color for each group)}

\item{prefix}{a character string defining a prefix of output file.}

\item{num.km.clusters}{a vector of manually defined number(s) of clusters.
By default, the values(s) are calculated automatically (a default value is
NULL).}
}
\value{
a data.frame with cluster information assigned to each sample;
a list of results from PCA;
PCA plot(s) saved into png.
}
\description{
A wrapper function for K-means clustering.
For data sets with >1000 features, multiple subsets (number of sets = number
of features/1000; rounded up to the higher integer) will be generated,
maximum is 10 subsets. A number of randomly selected features (e.g. genes)
in 1 subset is limited to 2000.
A number of clusters for each subset is automatically estimated using BIC and
a consensus (based on all the subsets) of best n k-means is returned and used
for clustering.
If K-means estimation using BIC fails, a number of clusters will be based on
the number of samples (2-6 clusters for data set with less than 100 samples;
2-11 for data sets with 101-500 samples, and 2-16 clusters for data sets with
more than 500 samples).
As a result, an information about the clusters assigned to each sample and
the details from PCA are provided.
The clusters are visualized on a PCA plot and saved into the png.
A summary describing the best model is printed on the screen during
calculation.
}
\examples{
\dontrun{
runKmeans(campp2_brca_1_batchCorrected, PCA.labels = "none", cols=NULL,
prefix="test", num.km.clusters=NULL)
}
}
