% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Kmeans.R
\name{runKmeans}
\alias{runKmeans}
\title{K-means clustering}
\usage{
runKmeans(
  data,
  num.subsets = "automatic",
  subset.size = "automatic",
  show.PCA.labels = FALSE,
  prefix,
  colors = NULL,
  num.km.clusters = NULL,
  seed = NULL,
  pca.scale = FALSE
)
}
\arguments{
\item{data}{a data.frame of feature (e.g. gene) counts}

\item{num.subsets}{a number of sub-sets to be generated for the estimation of
the number of clusters. This parameter has several options:
1) a numeric (integer) value estimated by the user.
2) a string "automatic" will let CAMPP2 estimate the number of the sub-sets
based on the number of features.
3) NULL option will use the whole set for the estimation
of the number of clusters. If using this option, subset.size must be also NULL.
Defaults is NULL.}

\item{subset.size}{a number of features included in each sub-set. This
parameter has several options:
1) a numeric (integer) value estimated by the user.
2) a string "automatic" will let CAMPP2 estimate the size of the sub-sets
based on the number of features.
3) NULL option will use all the features (no sub-sampling) for the estimation
of the number of clusters. If using this option, num.subsets must be also NULL.
Defaults is "automatic"}

\item{show.PCA.labels}{a boolean value (TRUE or FALSE) specifying if elements
(e.g. samples) should be labelled in the PCA plot including information about
the clusters. Labeling is based on column names of the input data.
Default value is "automatic".}

\item{prefix}{a character string defining a prefix of output file.}

\item{colors}{a vector of colors (one color for each cluster)}

\item{num.km.clusters}{either a vector of manually defined number(s) of
clusters, or NULL. Each number in this vector represent a plausible number
of clusters that the user would expected to be present in the data. If
multiple values provided, the function will automatically perform K-means
clustering using each of them as k (expected number of clusters) separately.
If this argument is NULL (default), optimal numbers of clusters are
calculated automatically based on the bayesian information criterion (mclust
package), applied to sub sampled data (see documentation for the whole
procedure).}

\item{seed}{a number for setting a random seed. If the argument is NULL,
the seed won't be set. Default is NULL.}

\item{pca.scale}{a boolean, if TRUE then data are scaled to unit variance
during PCA. Default is FALSE}
}
\value{
1) a list including:
   a) a data.frame with cluster information assigned to each sample;
   b) a PCA objects including data for generating PCA plots with cluster
   information (generated by PCA function from FactoMineR package);
2) 2D PCA plot(s) projecting samples (labeled with cluster number) over first 2
   principal components saved into .png.
3) plots of BIC values for each sample sub-set saved into .png file
}
\description{
A wrapper function for K-means clustering.
In a first step, a number of clusters is estimated.
By default, for data sets with >1000 features, multiple
subsets (number of sets = number of features/1000; rounded up to the higher
integer) will be generated, maximum is 10 sub-sets.
A number of randomly selected features (e.g. genes) in 1 subset is limited
to 2000. The values for number of subsets and size of the sub-sets might by
defined manually. The option of using the whole dataset for the estimation
of the number of the clusters is also available. By default, a number of
clusters for each sub-set is automatically estimated (using
mclustBIC function from mclust package) and a consensus (based on all the
subsets) of best n k-means is returned and used for K-means clustering.
If K-means estimation using BIC fails, a number of clusters will be based on
the number of samples (2-6 clusters for data set with less than 100 samples;
2-11 for data sets with 101-500 samples, and 2-16 clusters for data sets with
more than 500 samples). The number of clusters can also be supplied manually.
In the next step, K-means clustering using the information about the number
of clusters is done.
As results, an information about the clusters assigned to each sample and
results from PCA (FactoMineR) function are provided.
The clusters are visualized on PCA plot and saved into .png file.
Plot of BIC values for each of the sample sub-set is saved into .png file.
A summary describing the best model is printed on the screen during
calculation.
}
\examples{
\dontrun{
runKmeans(campp2_brca_1_batchCorrected[1:3000,], num.subsets= NULL,
subset.size=NULL, show.PCA.labels = FALSE, colors=NULL,
prefix="test", num.km.clusters=NULL, seed=123, pca.scale=FALSE)

}
}
